### 2019-05-08 #3 @lwasser
EarthPy: Software Submission for Review
Submitting Author: Leah Wasser (@lwasser)
Repository Link:  https://github.com/earthlab/earthpy
Version submitted:   0.7
Version accepted: TBD

---

```
Python is a generic programming language designed to support many different applications. Because of this, many commonly performed spatial tasks for science including plotting and working with spatial data take many steps of code. EarthPy takes advantage of functionality developed for raster data (rasterio) and vector data (geopandas) and simplifies the code needed to :

    Stack raster bands from data such as Landsat into an easy to use numpy array
    Work with masks to set bad pixels such a those covered by clouds and cloud-shadows to NA (mask_pixels())
    Plot rgb (color), color infrared and other 3 band combination images (plot_rgb())
    View histograms of sets of raster
    Create discrete (categorical) legends

EarthPy also has an io module that allows users to

    Quickly access pre-created datasubsets used in the earth-analytics courses hosted on www.earthdatascience.org
    Download other datasets that they may want to use in their workflows.

```

- Please indicate which [category or categories](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html) this package falls under:

\* Please fill out a pre-submission inquiry before submitting a data visualization package. For more info, see [this section](https://www.pyopensci.org/dev_guide/peer_review/aims_scope.html#notes-on-categories) of our guidebook.

- Explain how the and why the package falls under these categories (briefly, 1-2 sentences):
This package wraps around rasterio and geopandas to make working with geospatial data easier.

The target audience is people working with different types of raster and vector data in python. There are many operations that are often repeated by users that require a lot of code. This package simplifies these operations so users can quickly explore their data.

Not that we know of! this is why we created this package.

-   If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or @tag the editor you contacted:

For details about the pyOpenSci packaging requirements, see our [packaging guide](https://www.pyopensci.org/dev_guide/packaging/packaging_guide.html). Confirm each of the following by checking the box.  This package:

- [x] has an [OSI approved license](https://opensource.org/licenses)

- [ ] The package has an **obvious research application** according to JOSS's definition in their [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements). Be aware that completing the pyOpenSci review process **does not** guarantee acceptance to JOSS. Be sure to read their submission requirements (linked above) if you are interested in submitting to JOSS.
- [ ] The package is not a "minor utility" as defined by JOSS's [submission requirements](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements): "Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable." pyOpenSci welcomes these packages under "Data Retrieval", but JOSS has slightly different criteria.

- [x] I agree to abide by [pyOpenSci's Code of Conduct](https://www.pyopensci.org/dev_guide/peer_review/coc.html) during the review process and in maintaining my package should it be accepted.

**P.S.** *Have feedback/comments about our review process? Leave a comment [here](https://github.com/pyOpenSci/governance/issues/8)*

NOTE: I am actually not sure what research application means according to Joss!! may followup with Arfon on this.
