### 2019-05-08 earthlab/earthpy#318 @lwasser
EarthPy was just submitted to pyopensci!
https://github.com/pyOpenSci/software-review/issues/3

yay!
