### 2019-05-08 4a92ef4
Implemented changes suggested by Jenny. Add comments in the final section, made changes to the croppe loop so that it will work without previous rasters being opened, and the code is cleaner.
