+++ examples/plot_raster_stack_crop.py
@@ -9,13 +9,13 @@ Learn how to stack and crop satellite imagery with GeoPandas objects in EarthPy
 # Stacking and Cropping Rasters using EarthPy
 # ---------------------------------------------
 #
-# .. note:: The examples below will show you how to use the ``es.stack()`` and
-# ``es.crop_image()`` functions from EarthPy.
+# .. note::
+#       The examples below will show you how to use the ``es.stack()`` and
+#       ``es.crop_image()`` functions from EarthPy.
 
 ###############################################################################
 # Stacking Multi Band Imagery
 # -----------------------------
-#
 # Some remote sensing datasets are stored with each band in a separate file. However,
 # often you want to use all of the bands together in your analysis. For example if
 # you wish to plot multiple bands together to create a color image you will need all
@@ -42,19 +42,19 @@ from glob import glob
 import matplotlib.pyplot as plt
 import rasterio as rio
 from rasterio.plot import plotting_extent
-import earthpy as et
 import geopandas as gpd
+import earthpy as et
 import earthpy.spatial as es
 import earthpy.plot as ep
 
 
-###############################################################################
+########################################################################################
 # Get Example Data Ready for Stack
 # ----------------------------------
 # With EarthPy we can create a stack from all of the Landsat .tif files (one per band)
 # in a folder with the `es.stack()` function.
 
-#############################################################################
+###################################################################################
 # Error found on Windows systems
 # -------------------------------
 # .. note::
@@ -79,6 +79,8 @@ stack_band_paths.sort()
 output_dir = os.path.join("data", "outputs")
 if os.path.isdir(output_dir) == False:
     os.mkdir(output_dir)
+if os.path.isdir(os.path.join(output_dir + "outputraster")) == False:
+    os.mkdir(os.path.join(output_dir + "outputraster"))
 
 raster_out_path = os.path.join(output_dir, "raster.tiff")
 
@@ -120,7 +122,7 @@ ep.plot_rgb(
     stretch=True,
     extent=extent,
     str_clip=0.5,
-    title="Un-cropped Raster and Fire Boundary",
+    title="Un-cropped Raster",
 )
 plt.show()
 
@@ -153,7 +155,7 @@ ep.plot_rgb(
     stretch=True,
     extent=extent,
     str_clip=0.5,
-    title="Un-cropped Raster and Fire Boundary, No Data Value Selected",
+    title="Un-cropped Raster, No Data Value Selected",
 )
 plt.show()
 
@@ -171,28 +173,40 @@ plt.show()
 
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 
+# Getting data for the fire boundary we are going to clip the raster too.
+
 data_path_2 = et.data.get_data("cold-springs-fire")
+
+# Opening the boundary with GeoPandas.
 
 crop_bound = gpd.read_file(
     "data/cold-springs-fire/vector_layers/fire-boundary-geomac/co_cold_springs_20160711_2200_dd83.shp"
 )
 
-os.environ[
-    "PROJ_LIB"
-] = r"C:\Users\Nathan\Anaconda3\envs\earth-analytics-python\Library\share"
+# If you are on windows, make sure to set your environment here!
 
 # Change the Coordinate Reference System of the boundary to match that of the raster.
+# To do this, open one of the rasters are a RasterIO object, and get the profile.
+# You can use this profile to change the coordinate reference system of the boundary.
 
-crop_bound_utm13N = crop_bound.to_crs(raster_prof["crs"])
+with rio.open(stack_band_paths[0]) as raster_crs:
+    crop_raster_profile = raster_crs.profile
+    crop_bound_utm13N = crop_bound.to_crs(crop_raster_profile["crs"])
 
[…]
