+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+* A multi-point, polygon and line support to clip function (@lwasser, @nkorinek, #236)
 
 ## [0.6.8]
 * Add multi-panel plotting to plot_bands (@lwasser, #316)
