### 2019-05-10 4750f3a
Skip tests requiring internet on CI
