+++ earthpy/tests/test_io.py
@@ -87,6 +87,7 @@ eio.DATA_URLS["little-zip-file"] = [
 ]
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_urls_are_valid():
     """ Test responses for each dataset to ensure valid URLs. """
@@ -125,6 +126,7 @@ def test_invalid_dataset_key(eld):
         eld.get_data(key="some non-existent key")
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_valid_download_file(eld):
     """ Test that single files get downloaded. """
@@ -132,6 +134,7 @@ def test_valid_download_file(eld):
     assert os.path.isfile(file)
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_valid_download_zip(eld):
     """ Test that zipped files get downloaded and extracted. """
@@ -140,6 +143,7 @@ def test_valid_download_zip(eld):
     assert path_has_contents
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.parametrize("replace_arg_value", [True, False])
 @pytest.mark.vcr()
 def test_replace_arg_controle_overwrite(eld, replace_arg_value):
@@ -154,6 +158,7 @@ def test_replace_arg_controle_overwrite(eld, replace_arg_value):
         assert mtime1 == mtime2
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_arbitrary_url_file_download(eld):
     """ Verify that arbitrary URLs work for data file downloads. """
@@ -170,6 +175,7 @@ def test_invalid_data_type(eld):
         eld.get_data("invalid-data-type")
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_arbitrary_url_zip_download(eld):
     """ Verify that aribitrary URLs work for zip file downloads. """
@@ -180,6 +186,7 @@ def test_arbitrary_url_zip_download(eld):
     assert path_has_contents
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_url_download_tar_file(eld):
     """ Ensure that tar files are downloaded and extracted. """
@@ -187,6 +194,7 @@ def test_url_download_tar_file(eld):
     assert "abc.txt" in os.listdir(path)
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_url_download_tar_gz_file(eld):
     """ Ensure that tar.gz files are downloaded and extracted. """
@@ -194,6 +202,7 @@ def test_url_download_tar_gz_file(eld):
     assert "abc.txt" in os.listdir(path)
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.vcr()
 def test_url_download_txt_file_with_content_disposition(eld):
     """ Test arbitrary URL download with content-disposition. """
@@ -201,6 +210,7 @@ def test_url_download_txt_file_with_content_disposition(eld):
     assert path.endswith("abc.txt") and os.path.isfile(path)
 
 
+@pytest.mark.skipif(os.environ["CI"])
 @pytest.mark.parametrize("verbose_arg_value", [True, False])
 @pytest.mark.vcr()
 def test_verbose_arg_works(eld, verbose_arg_value, capsys):
