### 2019-05-10 a7dd1fb
Conditionally skip tests that require internet on CI builds
