+++ earthpy/tests/test_io.py
@@ -9,6 +9,12 @@ import geopandas as gpd
 import earthpy.io as eio
 
 
+RUNNING_ON_CI = False
+if "CI" in os.environ:
+    if os.environ["CI"]:
+        RUNNING_ON_CI = True
+
+
 @pytest.fixture
 def eld(tmpdir):
     return eio.Data(path=tmpdir)
@@ -87,7 +93,7 @@ eio.DATA_URLS["little-zip-file"] = [
 ]
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
 def test_urls_are_valid():
     """ Test responses for each dataset to ensure valid URLs. """
@@ -126,7 +132,7 @@ def test_invalid_dataset_key(eld):
         eld.get_data(key="some non-existent key")
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
 def test_valid_download_file(eld):
     """ Test that single files get downloaded. """
@@ -134,7 +140,7 @@ def test_valid_download_file(eld):
     assert os.path.isfile(file)
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
 def test_valid_download_zip(eld):
     """ Test that zipped files get downloaded and extracted. """
@@ -143,7 +149,7 @@ def test_valid_download_zip(eld):
     assert path_has_contents
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.parametrize("replace_arg_value", [True, False])
 @pytest.mark.vcr()
 def test_replace_arg_controle_overwrite(eld, replace_arg_value):
@@ -158,7 +164,7 @@ def test_replace_arg_controle_overwrite(eld, replace_arg_value):
         assert mtime1 == mtime2
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
 def test_arbitrary_url_file_download(eld):
     """ Verify that arbitrary URLs work for data file downloads. """
@@ -175,7 +181,7 @@ def test_invalid_data_type(eld):
         eld.get_data("invalid-data-type")
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
 def test_arbitrary_url_zip_download(eld):
     """ Verify that aribitrary URLs work for zip file downloads. """
@@ -186,7 +192,7 @@ def test_arbitrary_url_zip_download(eld):
     assert path_has_contents
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
 def test_url_download_tar_file(eld):
     """ Ensure that tar files are downloaded and extracted. """
@@ -194,7 +200,7 @@ def test_url_download_tar_file(eld):
     assert "abc.txt" in os.listdir(path)
 
 
-@pytest.mark.skipif(os.environ["CI"])
+@pytest.mark.skipif(RUNNING_ON_CI, reason="Fails intermittently on CI")
 @pytest.mark.vcr()
[…]
