### 2019-05-10 373d465
Define a skip_on_ci decorator
