+++ earthpy/tests/conftest.py
@@ -50,8 +50,3 @@ def basic_image_tif(tmpdir, basic_image):
 @pytest.fixture
 def image_array_2bands():
     return np.random.randint(10, size=(2, 4, 5))
-
-
-@pytest.fixture
-def one_band_3dims():
-    return np.random.randint(10, size=(1, 4, 5))
