### 2019-05-14 5d729f7
Took out the areas that were effected by the merge conflicts within the README.md.
