### 2019-05-14 398d8ee
Created a vignette to show a lot of the functionality found in the plot_bands function!
