### 2019-05-16 23a95e7
Made some needed changes to the readme and contributing files to help with building vignettes and finding examples. Also added changes in this pr to the changelog
