+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-* Add example vignette for stacking and cropping raster bands with EarthPy (@nkorinek, #307)
+* Add example vignette for stacking and cropping raster bands with EarthPy (@nkorinek, #267)
 
 ## [0.6.8]
 * Add multi-panel plotting to plot_bands (@lwasser, #316)
