### 2019-05-16 4153a57
updating from master (#322)
* start at updating the get started page

* black

* black

* doc update

* update from master and adding gallery

* Bump version: 0.6.3 → 0.6.4

* updated vignette

* black

* a few typos

* rearranging text

* fixing example output

* minor doc fix

* verbose false

* Remove download from setup.py

* added nodata support for stack

* added test for stack-nodata param

* added nodata param in doc

* Customizing docs generation

This commit improves on the default TOC handling of API documentation.
It introduces "Package Reference" as the top-level page for API docs
[…]
