+++ docs/conf.py
@@ -24,9 +24,9 @@ copyright = "2019, Earth Lab"
 author = "Earth Lab"
 
 # The short X.Y version
-version = ""
+version = "0.6.8"
 # The full version, including alpha/beta/rc tags
-release = ""
+release = "0.6.8"
 
 
 # -- General configuration ---------------------------------------------------
@@ -45,7 +45,16 @@ extensions = [
     "sphinx.ext.doctest",
     "sphinx.ext.napoleon",
     "sphinx.ext.viewcode",
+    "sphinx_gallery.gen_gallery",
 ]
+
+
+sphinx_gallery_conf = {
+    # path to your examples scripts
+    "examples_dirs": "../examples",
+    # path where to save gallery generated examples
+    "gallery_dirs": "gallery_vignettes",
+}
 
 # Show the source code for any plots in the documentation by default
 plot_include_source = True
@@ -77,6 +86,7 @@ exclude_patterns = [
     ".DS_Store",
     "earthpy/tests",
     "earthpy/example-data",
+    "api/earthpy.rst",
     "ignoreme",
 ]
 
