+++ CHANGELOG.md
@@ -5,8 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
 
+## [0.6.8]
+* Add multi-panel plotting to plot_bands (@lwasser, #316)
+* Fix bug in plot_rgb where multipanel plots are blank (@lwasser, #313)
+* Add example vignette for calculating and classifying NDVI with EarthPy (@jlpalomino, #266)
+
+## [0.6.7]
+* Add NoData masking support for `stack()` (@joemcglinchy, #282)
+* Fix multiline messages to use `"` vs `"""` (@lwasser, #270)
+
+## [0.6.6]
+* Add sphinx gallery for vignettes and update get started page (@lwasser #279, #203)
+* Add two example vignettes for using EarthPy with raster data (@lwasser)
+* Fix bug in `bytescale()` - ensure math to calc range is floating point vals (@lwasser #282)
+* Fix tests for `bytescale()` to ensure the bug raised in #282 is fixed / tested; also added comment to plot_rgb docs to ensure users consider nodata values before plotting (@lwasser #282)
+
+## [0.6.5] - 2019-03-37
 * Add continuous integration testing on osx via Travis CI (@mbjoseph #228)
 * Add cbar legend to `plot_bands()` and scaling parameters (@lwasser #274)
 * BUGFIX: `plot_bands()` doesn't plot single string titles properly + add test (@lwasser #258)
@@ -18,7 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added a code of conduct (@mbjoseph, #27)
 * Added CI testing across python versions and on Windows (@mbjoseph, #228)
 * Added tests for `EarthlabData` class (@mbjoseph, #37)
-
 
 ## [0.6.2] - 2019-02-19
 We have made significant changes in preparation for a 1.0 release
