+++ CONTRIBUTORS.rst
@@ -1,10 +1,21 @@
 Contributors
 ============
 
+Maintainers
+~~~~~~~~~~~~
+
+EarthPy is maintained by a core group of Python enthusiasts at Earth Lab - University of Colorado, Boulder
+
+* Leah Wasser (@lwasser)
+* Max Joseph (@mbjoseph)
+* Joseph McGlinchy (@joemcglinchy)
+* Jenny Palomino (@jlpalomino)
+
+
+Contributors
+~~~~~~~~~~~~
+
 We want to thank the following people for contributing to the ``earthpy``
 package.
 
-* Chris Holdgraf
-* Leah Wasser
-* Tim Head
-* Max Joseph
+* List to be added 

[…]
