### 2019-05-16 6f5b6e2
Added some functionality examples, also combined some examples together to make the vignette more effecient.
