### 2019-05-16 6bcea4b
Added the vignette in a link in the readme.
