### 2019-05-16 59ca616
Added a link to the sphinx gallery documentation help in the contributing file.
