+++ CONTRIBUTING.rst
@@ -175,7 +175,9 @@ format and format your pull request accordingly.
 To create a vignette for an EarthPy workflow, create a .py file that shows the
 necessary steps to complete the workflow. Make sure the file name begins with
 ``plot`` in order to ensure that the vignette is correctly built. Store the
-vignette in the ``examples`` folder within the ``earthpy`` directory.
+vignette in the ``examples`` folder within the ``earthpy`` directory. EarthPy
+uses Sphinx Gallery to build vignettes. Help for formatting and building
+vignettes can be found on `their website <https://sphinx-gallery.github.io>`_.
 
 
 To build the documentation, use the command::
