### 2019-05-20 ea87546
Add vignette landsat data to data_urls
