### 2019-05-20 3b263d7
Update URL for landsat vignette data and add to changelog
