### 2019-05-20 6beaacd
Update earthpy/tests/test_clip.py
Co-Authored-By: Max Joseph <maxwellbjoseph@gmail.com>
