### 2019-05-20 1e845cb
another small fix to clip implementaiton - any
