### 2019-05-20 3bfd684
Remove dup test fixture
