+++ earthpy/tests/test_plot_bands.py
@@ -17,11 +17,6 @@ def one_band_3dims():
     )
 
 
-@pytest.fixture
-def one_band_3dims():
-    return np.random.randint(10, size=(1, 4, 5))
-
-
 def test_arr_parameter():
     """Raise an AttributeError if an array is not provided."""
     with pytest.raises(
