### 2019-05-20 6b8cd46
Add new key for landsat vignette data
