+++ earthpy/tests/test_clip.py
@@ -8,8 +8,9 @@ import geopandas as gpd
 import earthpy.clip as cl
 
 
-def make_locs_gdf():
-    """ Create a dummy point GeoDataFrame. """
+@pytest.fixture
+def point_gdf():
+    """ Create a point GeoDataFrame. """
     pts = np.array([[2, 2], [3, 4], [9, 8], [-12, -15]])
     gdf = gpd.GeoDataFrame(
         [Point(xy) for xy in pts],
@@ -19,18 +20,24 @@ def make_locs_gdf():
     return gdf
 
 
-def make_single_rect_poly_gdf():
-    """ Bounding box polygon. """
+@pytest.fixture
+def single_rectangle_gdf():
+    """Create a single rectangle for clipping. """
     poly_inters = Polygon([(0, 0), (0, 10), (10, 10), (10, 0), (0, 0)])
     gdf = gpd.GeoDataFrame(
         [1], geometry=[poly_inters], crs={"init": "epsg:4326"}
     )
-    gdf["attr2"] = ["site-boundary"]
+    gdf["attr2"] = "site-boundary"
     return gdf
 
 
-def make_smaller_clip_rect_poly_gdf():
-    """ Bounding box polygon. """
+@pytest.fixture
+def larger_single_rectangle_gdf():
+    """Create a slightly larger rectangle for clipping.
+
+     The smaller single rectangle is used to test the edge case where slivers
+     are returned when you clip polygons. This fixture is larger which
+     eliminates the slivers in the clip return."""
     poly_inters = Polygon([(-5, -5), (-5, 15), (15, 15), (15, -5), (-5, -5)])
     gdf = gpd.GeoDataFrame(
         [1], geometry=[poly_inters], crs={"init": "epsg:4326"}
@@ -39,38 +46,26 @@ def make_smaller_clip_rect_poly_gdf():
     return gdf
 
 
-def make_locs_buff():
-    """ Buffer points to create multi poly. """
-    buffered_locations = make_locs_gdf()
-    buffered_locations["geometry"] = buffered_locations.buffer(4)
-    buffered_locations["type"] = "plot"
-    return buffered_locations
+@pytest.fixture
+def buffered_locations(point_gdf):
+    """Buffer points to create a multi-polygon. """
+    buffered_locs = point_gdf
+    buffered_locs["geometry"] = buffered_locs.buffer(4)
+    buffered_locs["type"] = "plot"
+    return buffered_locs
 
 
-def make_donut_geom():
-    """ Make a donut geometry. """
+@pytest.fixture
+def donut_geometry(buffered_locations, single_rectangle_gdf):
+    """ Make a geometry with a hole in the middle (a donut). """
     donut = gpd.overlay(
-        make_locs_buff(),
-        make_single_rect_poly_gdf(),
-        how="symmetric_difference",
+        buffered_locations, single_rectangle_gdf, how="symmetric_difference"
     )
     return donut
 
 
 @pytest.fixture
-def locs_gdf():
-    """ Create a sample point GeoDataFrame.
-
-    This fixture calls make_locs_gdf(), which is a function that is used in
-    multiple fixtures. But, fixtures are not supposed to be used like that:
-
-    see https://github.com/pytest-dev/pytest/issues/3950 for discussion
-    """
-    return make_locs_gdf()
-
-
-@pytest.fixture
-def linez_gdf():
+def two_line_gdf():
     """ Create Line Objects For Testing """
     linea = LineString([(1, 1), (2, 2), (3, 2), (5, 3)])
     lineb = LineString([(3, 4), (5, 7), (12, 2), (10, 5), (9, 7.5)])
[…]
