### 2019-05-21 220f237
Changed the data folder for this vignette.
