### 2019-05-21 2e8a9d9
Fixed a couple of filepaths I left out in the last commit.
