### 2019-05-21 c2441c7
Implemented new dataset for vignette examples.
