+++ CONTRIBUTING.rst
@@ -172,7 +172,7 @@ that our documentation files are in
 <http://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html>`_
 format and format your pull request accordingly.
 
-To create a vignette for an EarthPy workflow, create a .py file that shows the
+To create a vignette for an EarthPy workflow, create a ``.py`` file that shows the
 necessary steps to complete the workflow. Make sure the file name begins with
 ``plot`` in order to ensure that the vignette is correctly built. Store the
 vignette in the ``examples`` folder within the ``earthpy`` directory. EarthPy
