+++ examples/plot_raster_stack_crop.py
@@ -2,7 +2,7 @@
 Stack and Crop Raster Data Using EarthPy
 ========================================
 
-Learn how to stack and crop satellite imagery with GeoPandas objects in EarthPy
+Learn how to stack and crop satellite imagery using EarthPy
 """
 
 ###############################################################################
