+++ examples/plot_raster_stack_crop.py
@@ -113,6 +113,7 @@ extent = plotting_extent(array[0], raster_prof["transform"])
 # Plot Un-cropped Data
 # ------------------------------
 # You can see the boundary and the raster before the crop using `ep.plot_rgb()`
+# Notice that the data appear washed out.
 
 fig, ax = plt.subplots(figsize=(12, 12))
 ep.plot_rgb(
@@ -123,6 +124,18 @@ ep.plot_rgb(
     str_clip=0.5,
     title="RGB Image of Un-cropped Raster",
 )
+plt.show()
+
+
+################################################################################
+# Explore the Range of Values in the Data
+# ---------------------------------------
+# You can explore the range of values found in the data using the EarthPy ``hist()``
+# function. Do you notice any extreme values that may be impacting the stretch
+# of the image?
+
+fig, ax = plt.subplots(figsize=(12, 12))
+ep.hist(array_nodata)
 plt.show()
 
 ###########################################################################
@@ -145,7 +158,8 @@ extent_nodata = plotting_extent(
 ################################################################################
 # Plot Un-cropped Data
 # ------------------------------
-# You can see the boundary and the raster before the crop using ``ep.plot_rgb()``
+# Below, the un-cropped data are plotted using earthpy's ``ep.plot_rgb()`` function
+# Notice that the data appear washed out. 
 
 fig, ax = plt.subplots(figsize=(12, 12))
 ep.plot_rgb(
@@ -157,7 +171,6 @@ ep.plot_rgb(
     title="RGB image of Un-cropped Raster, No Data Value Selected",
 )
 plt.show()
-
 
 #############################################################################
 # Crop the data
@@ -180,17 +193,26 @@ crop_bound = gpd.read_file(
     "data/vignette-landsat/vector_layers/fire-boundary-geomac/co_cold_springs_20160711_2200_dd83.shp"
 )
 
-# If you are on windows, make sure to set your environment here!
-
-# Change the Coordinate Reference System of the boundary to match that of the raster.
-# To do this, open one of the rasters are a RasterIO object, and get the profile.
-# You can use this profile to change the coordinate reference system of the boundary.
+#############################################################################
+# Reproject the data
+# ------------------
+# .. note::
+#       If you are on windows, make sure to set your environment here! 
+#
+# The crop function won't work properly if the data are in different Coordinate
+# Reference Systems (CRS). To fix this, be sure to reproject the crop layer to match 
+# the CRS of your raster data.
+# To reproject your data, first get the CRS of the raster from the rasterio profile
+# object. Then use that to reproject using geopandas `.to_crs`` method.
 
 with rio.open(stack_band_paths[0]) as raster_crs:
     crop_raster_profile = raster_crs.profile
     crop_bound_utm13N = crop_bound.to_crs(crop_raster_profile["crs"])
 
-# This step crops all of the bands one by one. It uses enumerate to do dynamic naming
+#############################################################################
+# Crop Each Band
+# --------------
+# This step crops all of the bands one by one. It uses enumerate to implement dynamic naming
 # of raster outputs. The `i` variable will be an integer that counts up one with every
 # loop that is completed, and `bands` will be the raster bands in the folder.
 # It crops the rasters one at a time and writes them to an output directory.
@@ -202,7 +224,10 @@ for i, bands in enumerate(stack_band_paths):
         with rio.open(outpath, "w", **meta) as dest:
             dest.write(crop)
 
-# This section stacks the cropped rasters generated by the previous loop.
+#############################################################################
+# Stack All Bands
+# ---------------
+# Once the data are cropped, you are ready to create a new stack.
 
 stack = glob("data/outputs/outputraster*")
 cropped_array, array_raster = es.stack(stack, nodata=-9999)
