### 2019-05-21 c425a5a
Minor edits to fix some formatting and plotting issues.
