### 2019-05-21 dcc5116
Minor edits to fix some formatting and plotting issues.
