### 2019-05-21 #3 @luizirber
## Editor checks:

- [ ] **License:** The package has an OSI accepted license

---

## Editor comments

---

Reviewer: @carsonfarmer
Reviewer: @kysolvik
Due date: 2019-06-11
