### 2019-05-22 00f92e9
Update earthpy/clip.py
Co-Authored-By: Max Joseph <maxwellbjoseph@gmail.com>
