### 2019-05-22 3a20c2b
pinning versions for RTD
