+++ environment.yml
@@ -3,11 +3,11 @@ channels:
   - conda-forge
 dependencies:
   # Core scientific python
-  - numpy
-  - matplotlib
+  - numpy=>1.16.3
+  - matplotlib=>3.1.0
   - python=3.6
 
   # Geo stuff
-  - geopandas
-  - rasterio
-  - scikit-image
+  - geopandas=>0.5.0
+  - rasterio=>1.0.23
+  - scikit-image=>0.15.0
