+++ environment.yml
@@ -4,10 +4,10 @@ channels:
 dependencies:
   # Core scientific python
   - numpy=>1.16.3
-  - matplotlib=>3.1.0
+  - matplotlib=3.1.0
   - python=3.6
 
   # Geo stuff
-  - geopandas=>0.5.0
-  - rasterio=>1.0.23
-  - scikit-image=>0.15.0
+  - geopandas=0.5.0
+  - rasterio=1.0.23
+  - scikit-image=0.15.0
