+++ environment.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
 dependencies:
   # Core scientific python
-  - numpy=>1.16.3
+  - numpy=1.16.3
   - matplotlib=3.1.0
   - python=3.6
 
