### 2019-05-22 4d01ccb
cleaning up rtd build- fewer resources
