+++ readthedocs.yml
@@ -5,7 +5,7 @@ conda:
 python:
    version: 3.7
    install:
-      - requirements: dev-requirements.txt
+      - requirements: rtd-requirements.txt
       - method: pip
         path: .
 sphinx:
