### 2019-05-22 5dcc8c1
Add plot_rg vignette to changelog
