### 2019-05-22 6636881
Add URLs to ploy_rgb and ndvi vignettes to README
