### 2019-05-22 5f14d37
Resolve conflict in changelog release
