### 2019-05-22 0c26c32
Revert "Resolve conflict in changelog release"
This reverts commit 5f14d37d02374782c9d60b0f4eccbbc02042f99a.
