### 2019-05-22 da41a8e
Revert "Add plot_rg vignette to changelog"
This reverts commit 5dcc8c12ed5949bb7881033c390d8d16d0e1a24c.
