### 2019-05-22 5acf7d9
Revert "Add plot_rgb vignette with black format"
This reverts commit d1139a8f7052bd64ac109672d07735d4bb2654fe.
