### 2019-05-22 481523f
version cleanup and py 3.7
