+++ environment.yml
@@ -3,11 +3,11 @@ channels:
   - conda-forge
 dependencies:
   # Core scientific python
-  - numpy=1.16.3
-  - matplotlib=3.1.0
-  - python=3.6
+  - numpy
+  - matplotlib
+  - python=3.7
 
   # Geo stuff
-  - geopandas=0.5.0
-  - rasterio=1.0.23
-  - scikit-image=0.15.0
+  - geopandas
+  - rasterio
+  - scikit-image
