### 2019-05-23 5ff8d9f
Implementing the documentation changes suggested in the pull request.
