+++ examples/plot_bands_functionality.py
@@ -2,33 +2,31 @@
 Plot Bands of Satellite Imagery with EarthPy
 ==================================================================
 
-Learn how to plot each individual band from satellite imagery with
-EarthPy. This guide will demonstrate how to plot every band from a
-satellite, as well as individual bands. It will also go over some
-of the customization possible when plotting bands.
+Learn how to use the EarthPy ``plot_bands()`` function to quickly plot
+a single or many raster bands of an image. ``Plot_bands()`` can also
+be used to plot single raster layers with legends. Additionally,
+learn some of the customization possible when plotting bands.
 
 """
 
 ###############################################################################
-# Plotting Satellite Imagery Bands in Python Using EarthPy
-# ---------------------------------------------------------
+# Plot Raster Data Layers Using EarthPy
+# ------------------------------------------------------------------------
 #
 # .. note::
 #    The examples below will show you how to use the ``plot_bands()`` function
-#    to plot individual satellite bands in python. To plot rgb data, read help
-#    documentation related to `ep.plot_rgb()`.
+#    to plot individual raster layers in images using python. To plot rgb data,
+#    read help documentation related to ``ep.plot_rgb()``.
 #
-# The example below walks you through a typical workflow for plotting all bands
-# found in Landsat 8 data with EarthPy.
-#
-# First, you will create a stack of bands using Landsat 8 data and then
+# In this vignette, you will use Landsat 8. To begin, you will create a stack of
+# bands using Landsat 8 data and then plot the raster layers in various ways.
 
 ###############################################################################
 # Import Packages
 # ---------------
 #
-# To begin, import the needed packages. You will use a combination of several EarthPy
-# modules including spatial and plot.
+# In order to use the ``plot_bands()`` function with Landsat 8 data, the
+# following packages need to be imported.
 
 import os
 from glob import glob
@@ -37,9 +35,6 @@ import earthpy as et
 import earthpy.spatial as es
 import earthpy.plot as ep
 
-# Get data for example
-data = et.data.get_data("vignette-landsat")
-
 ###############################################################################
 # Import Example Data
 # -------------------
@@ -47,6 +42,9 @@ data = et.data.get_data("vignette-landsat")
 # To get started, make sure your directory is set. Then, create a stack from all of
 # the Landsat .tif files (one per band).
 
+# Get data for example
+data = et.data.get_data("vignette-landsat")
+# Setting home directory
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 
 # Stack the Landsat 8 bands
@@ -61,9 +59,9 @@ array_stack, meta_data = es.stack(landsat_path)
 # Plot All Bands in a Stack
 # --------------------------
 #
-# When `ep.plot_bands()` is handed a three dimensional numpy array, as is created
-# when `es.stack()` is ran, it will plot all layers of the numpy array. To title
-# all of the images individually, a list of titles must be submitted as an argument.
+# When you give ``ep.plot_bands()`` a three dimensional numpy array, as is created
+# when ``es.stack()`` is ran, it will plot all layers of the numpy array. To title
+# all of the images individually, you must submit a list of titles as an argument.
 # The list must contain the same number of strings as there are bands in the stack.
 
 titles = ["Ultra Blue", "Blue", "Green", "Red", "NIR", "SWIR 1", "SWIR 2"]
@@ -74,9 +72,9 @@ plt.show()
 # Plot One Band in a Stack
 # ------------------------
 #
-# When `ep.plot_bands` is handed a one dimensional numpy array, such as one layer
-# from the return of `es.stack`, it will just plot that band. If a colorbar is
-# not desired, it can be toggled off.
[…]
