+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
-* Add example vignette for plotting individual satellite bands with EarthPy (@nkorinek, #315)
+* Add vignette for using the plot_bands() function (@nkorinek, #315)
 * Add URL to io.py for Landsat vignette dataset (@jpalomino, #309)
 
 ## [0.6.8]
