### 2019-05-23 6d602fe
Update CHANGELOG and README with plot_rgb vignette
