### 2019-05-23 af7836b
Resolving merge conflicts with CHANGELOG.md
