### 2019-05-23 066664e
Implemented a few changes I missed in the README and the vignette. Minor documentation updates.
