### 2019-05-23 3eca439
Updated NDVI vignette to use vignette landsat data.
