### 2019-05-23 119b848
A few plot updates and minor edits!
