### 2019-05-23 1007a16
curious what happens when pip is called
