+++ docs/environment.yml
@@ -1,7 +1,7 @@
 name: earthpy_docs
 channels:
 - conda-forge
-dependencies:
+pip:
 - python=3.7
 #- numpy
 #- matplotlib=3.0.3
