+++ rtd-requirements.txt
@@ -1,5 +1,4 @@
 m2r=0.2.1
-pandas=0.24.2
 sphinx=1.8.0
 sphinx-autobuild=0.7.1
 sphinx_rtd_theme
