+++ docs/environment.yml
@@ -1,9 +1,11 @@
 name: earthpy_docs
 channels:
-- conda-forge
+  - conda-forge
 dependencies:
-- python=3.7
-- geopandas=0.4.1
-- rasterio=1.0.21
-- scikit-image
-- m2r=0.2.1
+  - python=3.7
+  - geopandas=0.4.1
+  - rasterio=1.0.21
+  - scikit-image
+
+  - pip:
+      - m2r=0.2.1
