### 2019-05-24 4fa0236
minor edits to the vignette
