### 2019-05-24 8692ce0
apply mask when stacking
