### 2019-05-24 e41ae15
fixing plot with mask of na values
