### 2019-05-24 dabd985
code fix - nodata is not a plot_bands arg
