### 2019-05-24 d6b8ec7
Resolve conflicts in CHANGELOG and README
