### 2019-05-24 9980a84
Add raster extent to plot_rgb vignette
