### 2019-05-28 a610b2b
First draft of the function. Includes all parameters requested, no tests have been made yet
