### 2019-06-04 ece1016
added code and tests for mismatched arrays, CRS, and Affine xform in stack
