### 2019-06-04 #3 @luizirber
Hi @kysolvik and @carsonfarmer! Anything I can do to help you with your reviews? Next steps are:

- Post this template in a comment here: https://www.pyopensci.org/dev_guide/appendices/templates#review-template
- Follow the template and https://www.pyopensci.org/dev_guide/peer_review/reviewer_guide.html for you review.
