### 2019-06-04 #3 @lwasser
our docs are building again too - yay! they were down due to a RTD issue. thank you @luizirber
