### 2019-06-06 #3 @lwasser
hey guys @carsonfarmer @luizirber  -- wanted to followup on here. @kysolvik pointed out that there is a potential conflict of interest here given Kylen has worked for me and i'm not sure if that would apply with @carsonfarmer as well given we were both a part of earth lab for some time. we may need to reassign reviewers. @luizirber i tried to ping you on slack but am not sure if you saw it! just want to ensure this review moves forward given we definitely need to swap out atleast one reviewer!! i was wondering if @leouieda could do it instead of @kysolvik  and kylen could do one of martin's submissions as one idea. but very open to thoughts.
