### 2019-06-06 1baff87
Minor changes to crop_all function, added tests and fixtures for conftest and test_crop_image
