### 2019-06-06 1c334d1
Added final tests requested for crop_all, added example to crop_all docstring.
