+++ earthpy/tests/test_crop_image.py
@@ -5,6 +5,7 @@ import numpy as np
 import pytest
 import geopandas as gpd
 import rasterio as rio
+from affine import Affine
 from shapely.geometry import Polygon, Point, LineString
 import earthpy.spatial as es
 
@@ -138,7 +139,7 @@ def test_crop_image_fails_empty_list(basic_image_tif, basic_geometry):
 
 
 def test_crop_all_returns_list(in_paths, output_dir, basic_geometry_gdf):
-    """Test that crop all returns a list"""
+    """Test that crop all returns a list. """
     img_list = es.crop_all(
         in_paths, output_dir, basic_geometry_gdf, overwrite=True
     )
@@ -148,23 +149,15 @@ def test_crop_all_returns_list(in_paths, output_dir, basic_geometry_gdf):
 def test_crop_all_returns_list_when_given_list(
     in_paths, output_file_list, basic_geometry_gdf
 ):
-    """Test that crop all returns a list"""
+    """Test that crop all returns a list when given a list instead of a directory. """
     img_list = es.crop_all(
         in_paths, output_file_list, basic_geometry_gdf, overwrite=True
     )
-    assert type(img_list) == list
-
-
-def test_crop_all_returns_list(in_paths, output_dir, basic_geometry_gdf):
-    """Test that crop all returns a list"""
-    img_list = es.crop_all(
-        in_paths, output_dir, basic_geometry_gdf, overwrite=True
-    )
-    assert len(img_list) == len(in_paths)
+    assert isinstance(img_list, list)
 
 
 def test_crop_all_files_exist(in_paths, output_dir, basic_geometry_gdf):
-    """Test that crop all returns a list"""
+    """Test that crop all actually creates the files in the directory. """
     img_list = es.crop_all(
         in_paths, output_dir, basic_geometry_gdf, overwrite=True
     )
@@ -172,22 +165,70 @@ def test_crop_all_files_exist(in_paths, output_dir, basic_geometry_gdf):
         assert os.path.exists(files)
 
 
-def test_crop_all_returns_list(in_paths, output_dir, basic_geometry_gdf):
-    """Test that crop all returns a list"""
-    with pytest.raises(ValueError):
+def test_crop_all_fails_overwrite(in_paths, output_dir, basic_geometry_gdf):
+    """Test that crop all fails when overwrite isn't set to True if files already exist. """
+    with pytest.raises(ValueError, match="The file "):
         es.crop_all(in_paths, output_dir, basic_geometry_gdf)
 
 
 def test_crop_all_fails_bad_dir(in_paths, basic_geometry_gdf):
-    """If user provides a bad directory path raise an error."""
+    """Test crop all fails if user provides a bad directory path. """
     bad_path = "Badpath"
-    with pytest.raises(ValueError):
-        es.crop_all(in_paths, bad_path, basic_geometry_gdf)
+    with pytest.raises(TypeError, match="The output directo"):
+        es.crop_all(in_paths, bad_path, basic_geometry_gdf, overwrite=True)
 
 
-def test_crop_all_returns_list(in_paths, output_dir, basic_geometry_gdf):
-    """Test that crop all returns a list"""
+def test_crop_all_fails_bad_dir_list(in_paths, basic_geometry_gdf):
+    """Test crop all fails if user provides a list of bad directory path. """
+    bad_path_list = ["Badpath0", "Badpath1", "Badpath2", "Badpath3"]
+    with pytest.raises(TypeError, match="The output directo"):
+        es.crop_all(
+            in_paths, bad_path_list, basic_geometry_gdf, overwrite=True
+        )
+
+
+def test_crop_all_returns_list_of_same_len(
+    in_paths, output_dir, basic_geometry_gdf
+):
[…]
