+++ earthpy/tests/test_stack.py
@@ -20,24 +20,24 @@ def in_paths(basic_image_tif):
 
 @pytest.fixture
 def in_paths_mismatch(basic_image_tif, basic_image_tif_2):
-    """ Input file paths for tifs of different sizes to stack. """
+    """ Input file paths for tifs of different dimensions to stack. """
     return [basic_image_tif] * 4 + [basic_image_tif_2]
 
 
 @pytest.fixture
 def in_paths_CRS_mismatch(basic_image_tif, basic_image_tif_CRS):
-    """ Input file paths for tifs of different sizes to stack. """
+    """ Input file paths for tifs of same dimension but different CRS to stack. """
     return [basic_image_tif] * 4 + [basic_image_tif_CRS]
 
 
 @pytest.fixture
 def in_paths_Affine_mismatch(basic_image_tif, basic_image_tif_Affine):
-    """ Input file paths for tifs of different sizes to stack. """
+    """ Input file paths for tifs of same dimension but different affine transform to stack. """
     return [basic_image_tif] * 4 + [basic_image_tif_Affine]
 
 
 def test_stack_array_size_mismatch(in_paths_mismatch):
-    """ Test for error raised when array sizes are not all equal. """
+    """ Test for error raised when array dimensions (nrows, ncols) are not all equal. """
 
     with pytest.raises(ValueError, match="same rows and columns"):
         es.stack(in_paths_mismatch)
