### 2019-06-06 c9ffbc3
First draft of hillshade vignette.
