### 2019-06-06 d0778e0
Tests failed on CI due to issue with how I was splitting files based on a period. With this fix hopefully it will fix the issues that CI found.
