### 2019-06-06 2489562
Added link to Readme and changes to changelog.
