### 2019-06-06 22691f5
Fixed a bug with the example data in the doc string that was causing CI to fail.
