### 2019-06-10 #3 @luizirber
Good point, I reset the reviewers and changed the label back to seeking reviewers.

If @leouieda agree to review this one, maybe @marskar would like to be a reviewer? It is a geospatial submission, but since @leouieda can cover the specific area I think it is good to have someone with a distinct expertise taking a look too.

@marskar @leouieda what do you think?
