+++ examples/plot_dem_hillshade.py
@@ -14,21 +14,26 @@ Learn how to create a hillshade Layer from a DEM using EarthPy
 #       function from EarthPy.
 
 ###############################################################################
-# Create hillshade from DEM
-# -----------------------------
-# Creating a hillshade layer can be a great way to help visualize elevation
-# data, and make your map more presentable. This vignette will show you
-# how to create a hillshade layer using EarthPy, and the different ways to
-# modify the layer to get the hillshade to look exactly how you want it to.
+# Create a Hillshade from a Digital Elevation Model (DEM)
+# -------------------------------------------------------
+# A hillshade is a 3D representation of a surface. Hillshades are generally
+# rendered in greyscale. The darker and lighter colors represent the shadows
+# and highlights that you would visually expect to see in a terrain model.
+# Hillshades are often used as an underlay in a map, to make the data appear
+# more 3-Dimensional and thus visually interesting. This vignette will show
+# you how to create a hillshade from a DEM using EarthPy. It will highlight
+# how to adjust the azimuth and other settings that will impact how the
+# shadows are modeled in the data.
 #
-# To begin using the EarthPy ``hillshade()`` function, import the needed
+# The hillshade function is a part of the spatial module in EarthPy.
+# To begin using the EarthPy hillshade() function, import the needed
 # packages and create an array to be plotted.
 
 ###############################################################################
 # Import Packages
-# ------------------------------
+# ----------------
 #
-# You will need several packages to plot hillshade. You will use RasterIO to
+# You will need several packages to plot hillshade. You will use Rasterio to
 # open up the DEM file you will use to create the hillshade layer. You will
 # primarily be using the EarthPy spatial module in this vignette.
 
@@ -40,92 +45,82 @@ import earthpy.spatial as es
 import earthpy.plot as ep
 import rasterio as rio
 
-# Downloading the data needed for this vignette
-
+# Download the data needed for this vignette
 data = et.data.get_data("colorado-flood")
 
 ####################################################################################
-# Opening up the DEM layer
+# Open up the DEM layer
 # -------------------------
-# To create a hillshade layer, we first need a DEM layer opened up as a numpy array.
-# We can do this with RasterIO. We will read in the elevation data, and take out any
-# data that is bad by setting values below 0 to numpy's not a number value. We then
+# To create a hillshade layer, you first need a DEM layer opened up as a numpy array.
+# You can do this with RasterIO. You will read in the elevation data, and take out any
+# data that is bad by setting values below 0 to numpy's not a number value. You then
 # plot the data with ``ep.plot_bands()`` to show that the data was read in properly.
 
-# Setting the home directory and getting the data for the exercise
-
+# Set the home directory and get the data for the exercise
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 dtm = "data/colorado-flood/spatial/boulder-leehill-rd/pre-flood/lidar/pre_DTM.tif"
 
-# Opening the DEM with RasterIO
-
+# Open the DEM with RasterIO
 with rio.open(dtm) as src:
     elevation = src.read(1)
-    # Fixing bad values
+    # Fix bad values
     elevation[elevation < 0] = np.nan
 
-# Plotting the data
-
+# Plott the data
 ep.plot_bands(elevation, scale=False, cmap="gist_earth")
 plt.show()
 
 ####################################################################################
-# Creating the hillshade layer
+# Create the hillshade layer
 # ----------------------------
 # Once the DEM numpy array is created, simply use ``es.hillshade()`` with the DEM
 # numpy array as an input to create the hillshade layer.
 
[…]
