### 2019-06-11 122414d
Fixed final merge conflicts
