### 2019-06-11 b507eba
Implemented changes suggest on GitHub to the vignette, lots of minor gramatical errors and formatting issues.
