### 2019-06-11 618b90e
Took away functionality for function that allowed the output_dir to be a list. This simplified the function significantly.
