### 2019-06-11 8961d6c
Fixed minor error in test functions.
