### 2019-06-11 cac305a
Rough draft for histogram vignette. Also made a minor edit to the plot_bands vignette, I saw a formatting issue.
