+++ examples/plot_bands_functionality.py
@@ -11,14 +11,14 @@ to plot a single raster layer with (or without) a legend.
 
 ###############################################################################
 # Plot Raster Data Layers Using EarthPy
-# ------------------------------------------------------------------------
+# --------------------------------------
 #
 # .. note::
 #    The examples below will show you how to use the ``plot_bands()`` function
 #    to plot individual raster layers in images using python. To plot rgb data,
 #    read help documentation related to ``ep.plot_rgb()``.
 #
-# In this vignette, you will use Landsat 8 data. To begin, you will create a 
+# In this vignette, you will use Landsat 8 data. To begin, you will create a
 # stack of
 # bands using Landsat 8 data. You will then plot the raster layers.
 
@@ -62,7 +62,7 @@ array_stack, meta_data = es.stack(landsat_path, nodata=-9999)
 # --------------------------
 #
 # When you give ``ep.plot_bands()`` a three dimensional numpy array,
-# it will plot all layers in the numpy array. You can create unique titles for 
+# it will plot all layers in the numpy array. You can create unique titles for
 # each image by providing a list of titles using the ``title=`` parameter.
 # The list must contain the same number of strings as there are bands in the stack.
 
