### 2019-06-11 4e2a1d5
Updated documentation to reflect changes to function.
