### 2019-06-13 acc9c02
Added new download folder to io file and used it for the hillshade vignette.
