### 2019-06-13 9fd171a
remove unneeded functions from shapely
