### 2019-06-13 4e3b37d
Added vignette to README and CHANGELOG.
