### 2019-06-13 6949cb6
Added week 13 twitter data to the io file in earthpy.
