### 2019-06-13 4024371
Added rst describing the data subsets available in EarthPy.
