+++ docs/earthpy-data-subsets.rst
@@ -13,6 +13,7 @@ Data Keys
 
 co-flood-extras
 ---------------
+
 Contains two additional csv files for the Colorado Flood exercise on earthdatascience.org.
 Includes precipitation and `temperature <https://figshare.com/articles/Week_2_earth_analytics/4502138/>`_
 data for the Boulder flood in 2013.
@@ -25,47 +26,57 @@ temperature, lidar, and clipping data for the 2013 Boulder Floods.More informati
 
 spatial-vector-lidar
 --------------------
+
 Main dataset for vector and lidar lessons on earthdatascience.org. Contains shapefiles for
 a variety of exercises, including road, county, and country shapefiles. Also contains
 lidar data from NEON for research done in California, as well as supporting data for the lidar data.
- More information `here<https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_.
+More information `here<https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_.
 
 cold-springs-modis-h5
 ---------------------
+
 Modis satellite data for the 2016 Cold Springs fire in h5 format. More information
 `here<https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
 
 cold-springs-fire
 -----------------
+
 Satellite and vector data for the 2016 Cold Springs fire. Includes Landsat, Modis, and NAIP
 data. More information `here<https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
 
 cs-test-naip
 ------------
+
 Naip imagery related to the 2016 Cold Springs fire.
 
 cs-test-landsat
 ---------------
+
 Landsat imagery related to the 2016 Cold Springs fire.
 
 ndvi-automation
 ---------------
+
 Landsat imagery and vector files for the automation assigment on earthdatascience.org
 The data covers sites studied by NEON. More information `here<https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_.
 
 vignette-landsat
 ----------------
+
 Landsat imagery used in the vignette `gallery<https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html>`_.
 
 vignette-elevation
 ------------------
+
 Elevation data used in the vignette `gallery<https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html>`_.
 
 california-rim-fire
 -------------------
+
 Data related to 2013 California Rim fire. Includes shapefiles and aerial imagery of the event.
 More information `here<https://figshare.com/articles/Earth_Analytics_Applications_Data_for_Rim_Fire_California_2013/7749548>`_.
 
 twitter-flood
 -------------
+
 Geolocated twitter data related to the 2013 Boulder floods. More information `here<https://figshare.com/articles/Earth_Analytics_Python_2017_Colorado_Floods_Twitter_data/5603413>`_.
