### 2019-06-13 44694c5
More formatting fixes
