+++ examples/plot_dem_hillshade.py
@@ -1,8 +1,9 @@
 """
-Create A Hillshade Layer Using EarthPy
-========================================
+Create A Hillshade From a Digital Elevation Model Using EarthPy
+===============================================================
 
-Learn how to create a hillshade Layer from a DEM using EarthPy
+Learn how to create a hillshade from a DEM using the EarthPy
+es.hillshade() function.
 """
 
 ###############################################################################
@@ -10,7 +11,7 @@ Learn how to create a hillshade Layer from a DEM using EarthPy
 # ---------------------------------------
 #
 # .. note::
-#       The examples below will show you how to use the ``es.hillshade()``
+#       The examples below show you how to use the ``es.hillshade()``
 #       function from EarthPy.
 
 ###############################################################################
@@ -26,16 +27,13 @@ Learn how to create a hillshade Layer from a DEM using EarthPy
 # shadows are modeled in the data.
 #
 # The hillshade function is a part of the spatial module in EarthPy.
-# To begin using the EarthPy hillshade() function, import the needed
-# packages and create an array to be plotted.
 
 ###############################################################################
 # Import Packages
 # ----------------
 #
-# You will need several packages to plot hillshade. You will use Rasterio to
-# open up the DEM file you will use to create the hillshade layer. You will
-# primarily be using the EarthPy spatial module in this vignette.
+# You will need several packages to get started. You will use Rasterio to
+# open up the DEM file needed to create the hillshade layer.
 
 import os
 import numpy as np
@@ -49,12 +47,10 @@ import rasterio as rio
 data = et.data.get_data("vignette-elevation")
 
 ####################################################################################
-# Open up the DEM layer
-# ----------------------
-# To create a hillshade layer, you first need a DEM layer opened up as a numpy array.
-# You can do this with RasterIO. You will read in the elevation data, and take out any
-# data that is bad by setting values below 0 to numpy's not a number value. You then
-# plot the data with ``ep.plot_bands()`` to show that the data was read in properly.
+# Open up the DEM
+# ---------------
+# To begin, open your DEM layer as a numpy array using Rasterio. Below you set all 
+# terrain values < 0 to ``nan``. Next, plot the data using ``ep.plot_bands()``.
 
 # Set the home directory and get the data for the exercise
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
@@ -71,20 +67,18 @@ ep.plot_bands(elevation, scale=False, cmap="gist_earth")
 plt.show()
 
 ####################################################################################
-# Create the hillshade layer
-# ----------------------------
-# Once the DEM numpy array is created, simply use ``es.hillshade()`` with the DEM
-# numpy array as an input to create the hillshade layer.
+# Create the Hillshade
+# --------------------
+# Once the DEM is read in, call ``es.hillshade()`` to create the hillshade.
 
-# Create the hillshade layer
+# Create and plot the hillshade with earthpy
 hillshade = es.hillshade(elevation)
 
-# Plot the hillshade layer
 ep.plot_bands(hillshade, scale=False, cbar=False)
 plt.show()
 
 ####################################################################################
-# Change the azimuth of the sun
+# Change the Azimuth of the Sun
 # -------------------------------
 # The angle that sun light hits the landscape, impacts the shadows and highlights
 # created on the landscape. You can adjust the azimuth values to adjust highlights
@@ -100,8 +94,8 @@ ep.plot_bands(hillshade_azimuth_210, scale=False, cbar=False)
 plt.show()
 
 ####################################################################################
-# Change the angle altitude of the sun
-# ---------------------------------------
+# Change the Angle Altitude of the Sun
+# -------------------------------------
[…]
