### 2019-06-18 1142cb7
Fixed some issues with formatting that were causing the build to fail.
