+++ docs/earthpy-data-subsets.rst
@@ -22,7 +22,7 @@ colorado-flood
 --------------
 Main dataset for the Colorado Flood exercise on earthdatascience.org. Contains precipitation,
 temperature, lidar, and clipping data for the 2013 Boulder Floods.More information
-available `here<https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_.
+available `here <https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_.
 
 spatial-vector-lidar
 --------------------
@@ -30,19 +30,19 @@ spatial-vector-lidar
 Main dataset for vector and lidar lessons on earthdatascience.org. Contains shapefiles for
 a variety of exercises, including road, county, and country shapefiles. Also contains
 lidar data from NEON for research done in California, as well as supporting data for the lidar data.
-More information `here<https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_.
+More information `here <https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_.
 
 cold-springs-modis-h5
 ---------------------
 
 Modis satellite data for the 2016 Cold Springs fire in h5 format. More information
-available `here<https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
+available `here <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
 
 cold-springs-fire
 -----------------
 
 Satellite and vector data for the 2016 Cold Springs fire. Includes Landsat, Modis, and NAIP
-data. More information `here<https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
+data. More information `here <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
 
 cs-test-naip
 ------------
@@ -58,25 +58,25 @@ ndvi-automation
 ---------------
 
 Landsat imagery and vector files for the automation assigment on earthdatascience.org
-The data covers sites studied by NEON. More information `here<https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_.
+The data covers sites studied by NEON. More information `here <https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_.
 
 vignette-landsat
 ----------------
 
-Landsat imagery used in the vignette `gallery<https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html>`_.
+Landsat imagery used in the vignette `gallery <https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html>`_.
 
 vignette-elevation
 ------------------
 
-Elevation data used in the vignette `gallery<https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html>`_.
+Elevation data used in the vignette `gallery <https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html>`_.
 
 california-rim-fire
 -------------------
 
 Data related to 2013 California Rim fire. Includes shapefiles and aerial imagery of the event.
-More information `here<https://figshare.com/articles/Earth_Analytics_Applications_Data_for_Rim_Fire_California_2013/7749548>`_.
[…]
