### 2019-06-18 f464077
More RST formatting fixes. RST is hard.
