+++ docs/earthpy-data-subsets.rst
@@ -20,29 +20,30 @@ data for the Boulder flood in 2013.
 
 colorado-flood
 --------------
-Main dataset for the Colorado Flood exercise on earthdatascience.org. Contains precipitation,
-temperature, lidar, and clipping data for the 2013 Boulder Floods.More information
-available `here <https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_.
+
+Main `dataset for the Colorado Flood <https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_
+exercise on https://www.earthdatascience.org. Contains precipitation,
+temperature, lidar, and clipping data for the 2013 Boulder Floods.
 
 spatial-vector-lidar
 --------------------
 
-Main dataset for vector and lidar lessons on earthdatascience.org. Contains shapefiles for
+Main `dataset for vector and lidar <https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_
+lessons on https://www.earthdatascience.org. Contains shapefiles for
 a variety of exercises, including road, county, and country shapefiles. Also contains
 lidar data from NEON for research done in California, as well as supporting data for the lidar data.
-More information `here <https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_.
 
 cold-springs-modis-h5
 ---------------------
 
-Modis satellite data for the 2016 Cold Springs fire in h5 format. More information
-available `here <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
+`Modis satellite data <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_
+for the 2016 Cold Springs fire in h5 format.
 
 cold-springs-fire
 -----------------
 
-Satellite and vector data for the 2016 Cold Springs fire. Includes Landsat, Modis, and NAIP
-data. More information `here <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
+`Satellite and vector data for the 2016 Cold Springs fire <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_.
+Includes Landsat, Modis, and NAIP data.
 
 cs-test-naip
 ------------
@@ -57,8 +58,8 @@ Landsat imagery related to the 2016 Cold Springs fire.
 ndvi-automation
 ---------------
 
-Landsat imagery and vector files for the automation assigment on earthdatascience.org
-The data covers sites studied by NEON. More information `here <https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_.
+`Landsat imagery and vector files <https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_
+for the automation assigment on https://www.earthdatascience.org. The data covers sites studied by NEON.
 
 vignette-landsat
 ----------------
@@ -73,10 +74,10 @@ Elevation data used in the vignette `gallery <https://earthpy.readthedocs.io/en/
 california-rim-fire
 -------------------
 
-Data related to 2013 California Rim fire. Includes shapefiles and aerial imagery of the event.
[…]
