### 2019-06-18 d2acec0
Added changes to the changelog.
