### 2019-06-18 0e9650b
Added document to index to include it in the toctree
