### 2019-06-18 04479b5
Added crop_image to ndvi vignette.
