### 2019-06-18 0e0db85
Added plot titles and changed plot sizes.
