### 2019-06-18 579066e
examples/plot_dem_hillshade.py
