### 2019-06-18 6180f4c
Added changes to the changelog
