### 2019-06-19 e9c312d
changelog update before bump version
