### 2019-06-19 ddb4a1f
Bump version: 0.6.9 → 0.6.10
