### 2019-06-19 6fea454
Bump version: 0.6.10 → 0.7.0
