### 2019-06-20 6270744
Fixing merge conflicts
