### 2019-06-20 9ddb809
Fixing merge conflicts
