### 2019-06-20 #3 @lwasser
@luizirber  i think we have confirmation as of today from @marskar and @leouieda that they are willing to work on reviewing earthpy!! i'll find someone else for nbless :)
