### 2019-06-20 197ce3b
Added single band cropping to this vignette.
