### 2019-06-25 69e58be
Updated the data rst with suggestions from Jenny. Expanded a lot of the descriptions and fixed grammatical issues.
