+++ docs/earthpy-data-subsets.rst
@@ -2,80 +2,109 @@
 EarthPy Data Subsets
 =====================
 
-EarthPy comes with the ability to download example datasets using
-``earthpy.data.get_data()``. These datasets are used in example documentation
-on how to use EarthPy, as well as in the EarthPy related lessons on
-earthdatascience.org. There are many keys that can be used to get data
-this way, all of which can be found below.
+EarthPy comes with the ability to download example datasets using the
+``earthpy.data.get_data()`` function. These datasets are used in example
+documentation on how to use EarthPy as well as in the EarthPy related
+lessons on https://www.earthdatascience.org, the open learning portal developed
+and maintained by Earth Lab at CU Boulder. To download these datasets, you can
+provide the appropriate data key to the ``earthpy.data.get.data()`` function
+for one of the datasets listed below
+(e.g.. ``earthpy.data.get_data("colorado-flood")``).
 
 Data Keys
 =========
 
-co-flood-extras
----------------
-
-Contains two additional csv files for the Colorado Flood exercise on earthdatascience.org.
-Includes precipitation and `temperature <https://figshare.com/articles/Week_2_earth_analytics/4502138>`_
-data for the Boulder flood in 2013.
-
 colorado-flood
 --------------
 
-Main `dataset for the Colorado Flood <https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_
-exercise on https://www.earthdatascience.org. Contains precipitation,
-temperature, lidar, and clipping data for the 2013 Boulder Floods.
+This dataset contains `data from the 2013 Colorado Flood <https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_.
+The files include river discharge data from Boulder Creek at North 75th St. Near Boulder,
+CO., collected by the U.S. Geological Survey(USGS) on the 5th of October, 2016. Additionally,
+there is daily precipitation data collected by the USGS near Boulder, CO., which was
+collected from 2003-2013. Finally, there are digital elevation models (DEMs) stored
+in this dataset, collected near Boulder. The data for these DEMs were acquired by the
+National Ecological Observatory Network(NEON) during fortuitous test flights before
+and after the 2013 Colorado Floods. Digital terrain models (DTMs) and digital surface
+models (DSMs) were generated by the NEON Airborne Observation Platform data processing
+team. Hillshades and the clip-extent were generated by Leah Wasser at Earth Lab - University
+of Colorado Boulder.
+
+co-flood-extras
+---------------
+
+This dataset contains two supplemental CSV files for the 2013 Colorado Flood
+data. The files include `precipitation <https://figshare.com/articles/CO_Flood_2013_-_Classroom_Lab/4295360>`_
+for August 21, 2013 to October 11, 2013 in Boulder, Colorado as well as an
+example of `temperature data <https://figshare.com/articles/Week_2_earth_analytics/4502138>`_
+that includes both null and nodata values for use in a data cleaning process.
 
 spatial-vector-lidar
 --------------------
 
-Main `dataset for vector and lidar <https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_
-lessons on https://www.earthdatascience.org. Contains shapefiles for
-a variety of exercises, including road, county, and country shapefiles. Also contains
[…]
