### 2019-06-25 2e02c11
Updated the data rst with suggestions from Jenny. Expanded a lot of the descriptions and fixed grammatical issues.
