### 2019-06-25 11322a9
Added changes suggested on github. Mostly formatting and documentation changes.
