### 2019-06-25 0092622
Fixed a minor bug with the documentation that was causing travis to fail.
