### 2019-06-26 462df23
minor edits
