### 2019-06-27 4b1835f
Added changes suggested on github to the documentation.
