+++ docs/earthpy-data-subsets.rst
@@ -18,12 +18,12 @@ colorado-flood
 --------------
 
 This dataset contains `data from the 2013 Colorado Flood <https://figshare.com/articles/Earth_Analytics_Python_Colorado_Flood_Teaching_Subset/6815945>`_.
-The files include river discharge data from Boulder Creek at North 75th St. Near Boulder,
-CO., collected by the U.S. Geological Survey(USGS) on the 5th of October, 2016. Additionally,
+The files include river discharge data from Boulder Creek at North 75th St. near Boulder,
+CO., collected by the U.S. Geological Survey (USGS) on the 5th of October, 2016. Additionally,
 there is daily precipitation data collected by the USGS near Boulder, CO., which was
 collected from 2003-2013. Finally, there are digital elevation models (DEMs) stored
 in this dataset, collected near Boulder. The data for these DEMs were acquired by the
-National Ecological Observatory Network(NEON) during fortuitous test flights before
+National Ecological Observatory Network (NEON) during fortuitous test flights before
 and after the 2013 Colorado Floods. Digital terrain models (DTMs) and digital surface
 models (DSMs) were generated by the NEON Airborne Observation Platform data processing
 team. Hillshades and the clip-extent were generated by Leah Wasser at Earth Lab - University
@@ -43,8 +43,8 @@ spatial-vector-lidar
 
 This dataset contains `vector and Lidar data <https://figshare.com/articles/Week_4_Earth_Analytics_teaching_data_subset_-_lidar_and_insitu_for_California/4620268>`_
 This includes a subset of spatial data for the California Madera County
-and NEON Soaproot Saddle (SOAP) and San Joaquin Experimental Range(SJER) sites
-in California. There is also some other general spatial boundary layers from
+and NEON Soaproot Saddle (SOAP) and San Joaquin Experimental Range (SJER) sites
+in California. There are also some other general spatial boundary layers from
 natural earth, including global roads, political boundaries, and populated
 places. Finally, there are NEON Lidar data and insitu measurements for
 SOAP and SJER sites.
@@ -55,7 +55,7 @@ cold-springs-modis-h5
 This dataset contains imagery from the Moderate Resolution Imaging Spectrometer
 `(MODIS)    <https://www.earthdatascience.org/courses/earth-analytics-python/multispectral-remote-sensing-modis/modis-remote-sensing-data-in-python/>`_
 maintained by NASA. `The satellite data <https://figshare.com/articles/Earth_Analytics_Cold_Springs_Fire_Remote_Sensing_Data/6083210>`_
-is for the 2016 Cold Springs fire near Nederland, CO. and is in h5 format.
+are for the 2016 Cold Springs fire near Nederland, CO. and are in h5 format.
 
 cold-springs-fire
 -----------------
@@ -81,7 +81,7 @@ Landsat imagery related to the 2016 Cold Springs fire near Nederland, CO.
 ndvi-automation
 ---------------
 
-This dataset includes `landsat imagery and vector files <https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_
+This dataset includes `Landsat imagery and vector files <https://figshare.com/articles/Earth_Analytics_Python_HOMEWORK_Landsat_8_2017_for_SJER_and_HARV_Download_from_Earth_Explorer/7272500>`_
 for the SJER and SOAP NEON sites.
 
 vignette-landsat
@@ -109,4 +109,4 @@ MODIS Active Fire, MODIS Burned Area, and Incident Status Summary
 twitter-flood
 -------------
 
[…]
