### 2019-06-28 aecd1bc
Add Data description text to documentation (et.get_data support) (#371)
* Added rst describing the data subsets available in EarthPy.

* Fixed minor formatting issue.

* More formatting fixes

* Fixed some issues with formatting that were causing the build to fail.

* More RST formatting fixes. RST is hard.

* Added changes to the changelog.

* Added document to index to include it in the toctree

* Fixed broken link.

* Fixing merge conflicts

* Fixing merge conflicts

* Fixing merge conflicts

* Fixing merge conflicts

* Updated the data rst with suggestions from Jenny. Expanded a lot of the
[…]
