### 2019-06-28 441de92
Histogram vignette (#366)
* Rough draft for histogram vignette. Also made a minor edit to the plot_bands vignette, I saw a formatting issue.

* Added vignette to README and CHANGELOG.

* Added changes suggested on github. Mostly formatting and documentation changes.

* Fixed a minor bug with the documentation that was causing travis to fail.

* minor edit
