+++ examples/plot_bands_functionality.py
@@ -11,16 +11,15 @@ to plot a single raster layer with (or without) a legend.
 
 ###############################################################################
 # Plot Raster Data Layers Using EarthPy
-# ------------------------------------------------------------------------
+# --------------------------------------
 #
 # .. note::
 #    The examples below will show you how to use the ``plot_bands()`` function
 #    to plot individual raster layers in images using python. To plot rgb data,
 #    read help documentation related to ``ep.plot_rgb()``.
 #
-# In this vignette, you will use Landsat 8 data. To begin, you will create a 
-# stack of
-# bands using Landsat 8 data. You will then plot the raster layers.
+# In this vignette, you will use Landsat 8 data. To begin, you will create a
+# stack of bands using Landsat 8 data. You will then plot the raster layers.
 
 ###############################################################################
 # Import Packages
@@ -45,6 +44,7 @@ import earthpy.plot as ep
 
 # Get data for example
 data = et.data.get_data("vignette-landsat")
+
 # Set working directory
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 
@@ -62,11 +62,12 @@ array_stack, meta_data = es.stack(landsat_path, nodata=-9999)
 # --------------------------
 #
 # When you give ``ep.plot_bands()`` a three dimensional numpy array,
-# it will plot all layers in the numpy array. You can create unique titles for 
+# it will plot all layers in the numpy array. You can create unique titles for
 # each image by providing a list of titles using the ``title=`` parameter.
 # The list must contain the same number of strings as there are bands in the stack.
 
 titles = ["Ultra Blue", "Blue", "Green", "Red", "NIR", "SWIR 1", "SWIR 2"]
+# sphinx_gallery_thumbnail_number = 5
 ep.plot_bands(array_stack, title=titles)
 plt.show()
 
