+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+* Add vignette for using the hist() function (@nkorinek, #331)
 * Added file explaining all of the dataset available through EarthPy (@nkorinek, #369)
 * Added `es.crop_image()` example to the stack/crop vignette (@nkorinek, #368)
 
