### 2019-07-03 #3 @luizirber
Hi @leouieda and @marskar! You've been assigned, next steps are:

- Post this template in a comment here: https://www.pyopensci.org/dev_guide/appendices/templates#review-template
- Follow the template and https://www.pyopensci.org/dev_guide/peer_review/reviewer_guide.html for you review.
