### 2019-07-11 547d263
Added changes to the changelog and added the vignette to the README file
