+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.2]
+* Fix to the colorado-flood dataset (@lwasser, no associated issue). 
+  This is a result of some inconsistencies in our lessons. Minor changes to file names
+  have been made to address linux / mac/ windows case sensitivity issues.
+
 ## [0.7.1]
 * Updated colorado-flood dataset to fix file name (@lwasser, no associated issue)
 * Add vignette for using the hist() function (@nkorinek, #331)
