### 2019-07-25 4ee109a
Added in fix to earthpy.plot. Still need to add in tests.
