### 2019-07-30 21b60d9
Fixed issue that was causing build to fail
