### 2019-07-30 8499685
Added a few tests to earthpy for the new functionality of hist.
