+++ earthpy/tests/test_plot.py
@@ -133,3 +133,9 @@ def test_hist_plot_1_band_array(basic_image):
     f, ax = ep.hist(basic_image)
     assert len(f.axes) == 1
     plt.close(f)
+
+
+def test_hist_plot_2_band_array(image_array_2bands):
+    f, ax = ep.hist(image_array_2bands)
+    assert len(f.axes) == 2
+    plt.close(f)
