### 2019-07-30 310a8b2
Trying to increase coverage of hist function again.
