+++ earthpy/plot.py
@@ -401,7 +401,14 @@ def plot_rgb(
 
 
 def hist(
-    arr, colors=["purple"], figsize=(12, 12), cols=2, bins=20, title=None
+    arr,
+    colors=["purple"],
+    figsize=(12, 12),
+    cols=2,
+    bins=20,
+    title=None,
+    xlab="",
+    ylab="",
 ):
     """Plot histogram for each layer in a numpy array.
 
@@ -418,9 +425,13 @@ def hist(
         The number of columns for plot grid.
     bins : int (default = 20)
         The number of bins to generate for the histogram.
-    title : list (optional)
+    title : str or list (optional)
         A list of title values that should either equal the number of bands
-        or be empty.
+        or be empty. A string is accepted for a single dimension array.
+    xlab : str (optional)
+        The text to print on the x axis
+    ylab : str (optional)
+        The text to print on the y axis
     Returns
     ----------
     tuple
@@ -449,18 +460,24 @@ def hist(
         (<Figure size 800x300 with 3 Axes>, ...)
     """
 
+    if title:
+        if isinstance(title, str):
+            title = [title]
+
     # If the array is 3 dimensional setup grid plotting
-    if np.ma.is_masked(arr):
-        arrlis = []
-        for i in range(arr.shape[0]):
-            arrlis.append([arr[i][~arr[i].mask].data])
-        arr = np.array(arrlis)
     if arr.ndim > 2:
+        # Compress the arr if it's masked
+        if np.ma.is_masked(arr):
+            arrlis = []
+            for i in range(arr.shape[0]):
+                # before this was a list each time ... list of array?
+                arrlis.append(arr[i].compressed())
+            arr = np.array(arrlis)
+
         n_layers = arr.shape[0]
         if title and not len(title) == n_layers:
             raise ValueError(
-                "The number of plot titles should be the "
-                "same as the number of raster layers in your array."
+                "The number of plot titles should be the same as the number of raster layers in your array."
             )
         # Calculate the total rows that will be required to plot each band
         plot_rows = int(np.ceil(arr.shape[0] / cols))
@@ -482,19 +499,30 @@ def hist(
             ax.set_axis_off()
 
         return fig, axs
+
     elif arr.ndim <= 2:
+        # Test that only one title is provided for a 2-dim array
+        if title:
+            if len(title) > 1:
+                raise ValueError(
+                    "You have one array to plot but more than one title. Please provide a single title value."
+                )
+
         # Plot all bands
-        if arr.ndim == 2:
-            arr = arr.ravel()
+        if np.ma.is_masked(arr):
+            arr_comp = arr.compressed()
+        else:
+            arr_comp = arr.ravel()
+
         fig, ax = plt.subplots(figsize=figsize)
         ax.hist(
-            arr,
-            range=[np.nanmin(arr), np.nanmax(arr)],
+            arr_comp,
+            range=[np.nanmin(arr_comp), np.nanmax(arr_comp)],
             bins=bins,
             color=colors[0],
         )
         if title:
-            ax.set(title=title[0])
+            ax.set(title=title[0], xlabel=xlab, ylabel=ylab)
         return fig, ax
 
 
