### 2019-08-01 fbc338e
Made plot accept a color as a string
