### 2019-08-01 368d839
Added in the ability to modify the range variable of the histogram.
