### 2019-08-01 5e616d8
Adding tests for the range input for hist.
