### 2019-08-06 #3 @lwasser
hello there @luizirber @leouieda @marskar just checking in to see where this review is now?? i've bumped the version of earthpy up a few times since submitting so if the review hasn't started i might update the version here. and if it has and you guys have questions, please say the word!
