### 2019-08-06 2bd520d
Committing changes to show where issues in code are.
