### 2019-08-06 30189ea
Added gramatical and documentation suggestions from GitHub.
