+++ examples/plot_clip.py
@@ -11,16 +11,24 @@ polygon geometry using EarthPy.
 # Clip Vector Data in Python Using EarthPy
 # ---------------------------------------------
 #
+# .. note::
+#    The example below will show you how to use the clip_shp() function to clip
+#    vector data such as points, lines and polygons to a vector boundary.
+#
 # The example below walks you through a typical workflow for clipping one
 # vector data file to the shape of another. Both vector data files must be
 # opened with GeoPandas as GeoDataFrames and be in the same Coordinate
 # Reference System (CRS) for the ``clip_shp()`` function in EarthPy to work.
-# The object to be clipped will be clipped to the full extent of the clip
-# object.
+#
+# This example uses a line shapefile containing all major roads in North America
+# and a polygon shapefile containing the United States boundary to show you how
+# to prepare and clip vector data. In this example, the roads will be clipped
+# to the United States boundary.
 #
 # .. note::
-#    If there are multiple polygons in clip object, data from shp will be clipped
-#    to the total boundary of all polygons in clip object.
+#    The object to be clipped will be clipped to the full extent of the clip
+#    object. If there are multiple polygons in clip object, data from shp will
+#    be clipped to the total boundary of all polygons in clip object.
 
 ###############################################################################
 # Import Packages
@@ -41,7 +49,7 @@ import earthpy.clip as ec
 #
 # Once the packages have been imported, download the data needed for this
 # example: one line shapefile containing all major roads in North America,
-# and one polygon shapefile containing the United States border.
+# and one polygon shapefile containing the United States boundary.
 
 data = et.data.get_data("spatial-vector-lidar")
 
@@ -49,31 +57,44 @@ data = et.data.get_data("spatial-vector-lidar")
 # Open Files with GeoPandas and Reproject the Data
 # -------------------------------------------------
 #
-# The first step to clipping geometries with EarthPy is to open them with
-# GeoPandas as GeoDataFrames. To easily access both files, it is best to set
-# your home environment. Once they are open, you must ensure they are in the
-# same CRS, as ``clip_shp()`` won't work if there is a discrepancy in
-# projections between the two objects. If there is a discrepancy, you can
-# fix it with the function ``to_crs()`` from GeoPandas, as shown below.
+# Start by setting your working directory. Then, import the data files to
+# GeoDataFrames using GeoPandas.
+#
+# Recall that the data must be in the same CRS in order to use the
+# ``clip_shp()`` function. If the data are not in the same CRS, be sure to use
+# the ``to_crs()`` function from GeoPandas to match the projects between the
+# two objects, as shown below.
 
 # Set your home environment
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 
 # Open both files with GeoPandas
 roads = gpd.read_file(
-    r"data/spatial-vector-lidar/global/ne_10m_roads/ne_10m_n_america_roads.shp"
+    os.path.join(
+        "data",
+        "spatial-vector-lidar",
+        "global",
+        "ne_10m_roads",
+        "ne_10m_n_america_roads.shp",
+    )
 )
 country_boundary = gpd.read_file(
-    r"data/spatial-vector-lidar/usa/usa-boundary-dissolved.shp"
+    os.path.join(
+        "data", "spatial-vector-lidar", "usa", "usa-boundary-dissolved.shp"
+    )
 )
 
 # Reproject the roads layer to match the US boundary CRS
 roads = roads.to_crs(country_boundary.crs)
 
 # Plot the unclipped data
-# The plot below shows the data before it has been clipped
+# The plot below shows the roads data before it has been clipped
 fig, ax = plt.subplots(figsize=(12, 8))
 roads.plot(ax=ax)
+
[…]
