### 2019-08-06 059e882
Added tests for fully and partially masked arrays, as well as xlabel and ylabel input. Also fixed function so that xlabel and ylabel were assigned to the axes.
