### 2019-08-07 #3 @marskar
Hi @lwasser,
Let's work with the latest version.
I have a bio (not geo) background, so maybe @leouieda could address the science behind earthpy first and I will read his review and then follow up with more general packaging advice (testing, docs, etc.). Does that sound like a good plan?
