+++ examples/plot_clip.py
@@ -2,8 +2,8 @@
 Clip Vector Data with EarthPy
 ==================================================================
 
-Learn how to clip point, line, or polygon geometries with a
-polygon geometry using EarthPy.
+Learn how to clip point, line, or polygon geometries to the country_boundary
+of a polygon geometry using EarthPy.
 
 """
 
@@ -27,7 +27,7 @@ polygon geometry using EarthPy.
 #
 # .. note::
 #    The object to be clipped will be clipped to the full extent of the clip
-#    object. If there are multiple polygons in clip object, data from shp will
+#    object. If there are multiple polygons in clip object, the input data will
 #    be clipped to the total boundary of all polygons in clip object.
 
 ###############################################################################
@@ -69,32 +69,31 @@ data = et.data.get_data("spatial-vector-lidar")
 os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
 
 # Open both files with GeoPandas
-roads = gpd.read_file(
-    os.path.join(
-        "data",
-        "spatial-vector-lidar",
-        "global",
-        "ne_10m_roads",
-        "ne_10m_n_america_roads.shp",
-    )
+road_path = os.path.join(
+    "data",
+    "spatial-vector-lidar",
+    "global",
+    "ne_10m_roads",
+    "ne_10m_n_america_roads.shp",
 )
-country_boundary = gpd.read_file(
-    os.path.join(
-        "data", "spatial-vector-lidar", "usa", "usa-boundary-dissolved.shp"
-    )
+roads = gpd.read_file(road_path)
+
+country_path = os.path.join(
+    "data", "spatial-vector-lidar", "usa", "usa-boundary-dissolved.shp"
 )
+country_boundary = gpd.read_file(country_path)
 
 # Reproject the roads layer to match the US boundary CRS
 roads = roads.to_crs(country_boundary.crs)
 
-# Plot the unclipped data
-# The plot below shows the roads data before it has been clipped
+###############################################################################
+# The plot below shows the roads data before it has been clipped. Notice that
+# the ``.boundary`` method for a GeoPandas object is used to plot the
+# boundary rather than the filled polygon. This allows for other data, such as
+# the roads data, to be overlayed on top of the polygon boundary.
+
 fig, ax = plt.subplots(figsize=(12, 8))
 roads.plot(ax=ax)
-
-# Use the .boundary option in GeoPandas plotting to just plot the boundary of a polygon, instead of
-# a filled in version of the polygon. This allows for other data, such as the roads data, to be seen
-# in reference to the polygon boundary.
 country_boundary.boundary.plot(ax=ax, color="red")
 ax.set_title("Major NA Roads Unclipped to US Border", fontsize=20)
 ax.set_axis_off()
@@ -106,7 +105,9 @@ plt.show()
 #
 # Now that the data are opened as GeoDataFrame objects and in the same
 # projection, the data can be clipped! Recall that in this example, the roads
-# will be clipped to the United States boundary. To clip the data, make
+# will be clipped to the United States boundary.
+#
+# To clip the data, make
 # sure you put the object to be clipped as the first argument in
 # ``clip_shp()``, followed by the vector object (boundary) to which you want
 # the first object clipped. The function will return the clipped GeoDataFrame
