### 2019-08-08 7dc87b8
Add Clip vignette to earthpy (#400)
* Added changes to the changelog and added the vignette to the README file

* Changed dataset to be the correct dataset!

* Added gramatical and documentation suggestions from GitHub.

* Updated Changelog

* Commiting so I can merge

* black

* color changes
