+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+* Add vignette for using the clip_shp() function (@nkorinek, #378)
+
 ## [0.7.2]
 * Fix to the colorado-flood dataset which had nested folders (@lwasser, #387)
 
