### 2019-08-12 2efefed
Update hist function to support masked arrays and single dim arrays (#402)
* Added in fix to earthpy.plot. Still need to add in tests.

* Fixed issue that was causing build to fail

* Fixed a typo, oops!

* Added a few tests to earthpy for the new functionality of hist.

* Added another test to increase coverage

* Trying to increase coverage of hist function again.

* black

* black

* black

* black

* xlabel should follow mpl convention

* black

* black

* black

* black

* black (#395)

* black

* Made plot accept a color as a string

* Added in the ability to modify the range variable of
[…]
