+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.7.3]
 * Update `hist()` function to support masked arrays and single-dim arrays (@nkorinek, @lwasser #390)
 * Add vignette for using the `clip_shp()` function (@nkorinek, #378)
 
