+++ docs/conf.py
@@ -24,9 +24,9 @@ copyright = "2019, Earth Lab"
 author = "Earth Lab"
 
 # The short X.Y version
-version = "0.7.2"
+version = "0.7.3"
 # The full version, including alpha/beta/rc tags
-release = "0.7.2"
+release = "0.7.3"
 
 
 # -- General configuration ---------------------------------------------------
