### 2019-08-19 06c348e
update vignette to show plot image thumbnail
