### 2019-08-22 c28e9dc
Fixed the last two vignettes so that their thumbnails work.
