### 2019-08-29 bb40314
Added alpha arguements to plot_bands and hist
