### 2019-08-29 9b59670
Modified clip vignette to not rely on earthpy data
