+++ examples/plot_clip.py
@@ -20,10 +20,8 @@ of a polygon geometry using EarthPy.
 # opened with GeoPandas as GeoDataFrames and be in the same Coordinate
 # Reference System (CRS) for the ``clip_shp()`` function in EarthPy to work.
 #
-# This example uses a line shapefile containing all major roads in North America
-# and a polygon shapefile containing the United States boundary to show you how
-# to prepare and clip vector data. In this example, the roads will be clipped
-# to the United States boundary.
+# This example uses Polygons, a Line, and Points made with shapely and then
+# turned into GeoDataframes.
 #
 # .. note::
 #    The object to be clipped will be clipped to the full extent of the clip
@@ -37,65 +35,62 @@ of a polygon geometry using EarthPy.
 # To begin, import the needed packages. You will primarily use EarthPy's clip
 # utility alongside GeoPandas.
 
-import os
+import numpy as np
 import matplotlib.pyplot as plt
 import geopandas as gpd
-import earthpy as et
+from shapely.geometry import Polygon, LineString, Point
 import earthpy.clip as ec
 
 ###############################################################################
-# Import Example Data
+# Create Example Data
 # -------------------
 #
-# Once the packages have been imported, download the data needed for this
-# example: one line shapefile containing all major roads in North America,
-# and one polygon shapefile containing the United States boundary.
+# Once the packages have been imported, you need to create the shapes to clip.
+# You need to make two polygons, one line, and one point feature with shapely,
+# and then open those shapes up with GeoPandas.
 
-data = et.data.get_data("spatial-vector-lidar")
+polygon1 = Polygon([(0, 0), (0, 10), (10, 10), (10, 0), (0, 0)])
+polygon2 = Polygon([(-5, -5), (-5, 5), (5, 5), (5, -5), (-5, -5)])
+line = LineString([(3, 4), (5, 7), (12, 2), (10, 5), (9, 7.5)])
+pts = np.array([[2, 2], [3, 4], [9, 8]])
 
 ###############################################################################
 # Open Files with GeoPandas and Reproject the Data
 # -------------------------------------------------
 #
-# Start by setting your working directory. Then, import the data files to
-# GeoDataFrames using GeoPandas.
+# Open the data files to as GeoDataFrames using GeoPandas.
 #
-# Recall that the data must be in the same CRS in order to use the
-# ``clip_shp()`` function. If the data are not in the same CRS, be sure to use
-# the ``to_crs()`` function from GeoPandas to match the projects between the
-# two objects, as shown below.
+# .. note::
+#    Recall that the data must be in the same CRS in order to use the
+#    ``clip_shp()`` function. If the data are not in the same CRS, be sure to use
+#    the ``to_crs()`` function from GeoPandas to match the projects between the
+#    two objects, as shown below. In this example, since you make all of the data,
+#    you don't have to change the CRS.
 
-# Set your home environment
-os.chdir(os.path.join(et.io.HOME, "earth-analytics"))
-
-# Open both files with GeoPandas
-road_path = os.path.join(
-    "data",
-    "spatial-vector-lidar",
-    "global",
-    "ne_10m_roads",
-    "ne_10m_n_america_roads.shp",
+# Now that since you have all of the shapes created, you can open them with GeoPandas
+poly_gdf1 = gpd.GeoDataFrame(
+    [1], geometry=[polygon1], crs={"init": "epsg:4326"}
 )
-roads = gpd.read_file(road_path)
-
-country_path = os.path.join(
[…]
