### 2019-08-30 b322cd6
Add alpha arguments (#412)
* Added alpha arguements to plot_bands and hist
* Added issue numbers to the change
