### 2019-09-12 2e42a60
Added alpha arguement to hist and plot_bands, and a norm arguement to plot_bands, as well as tests for both.
