+++ earthpy/plot.py
@@ -82,6 +82,7 @@ def _plot_image(
     vmax=None,
     ax=None,
     alpha=1,
+    norm=None
 ):
 
     """
@@ -110,7 +111,11 @@ def _plot_image(
     alpha : float (optional)
         The alpha value for the plot. This will help adjust the transparency of the
         plot to the desired level.
-
+    norm : matplotlib Normalize object (Optional)
+        The normalized boundaries for custom values coloring. NOTE: For this arguement
+        to work, the scale arguement MUST be set to false. Otherwise, the values will
+        just be scaled from 0-255
+        
     Returns
     ----------
     ax : axes object
@@ -121,7 +126,7 @@ def _plot_image(
         arr_im = es.bytescale(arr_im)
 
     im = ax.imshow(
-        arr_im, cmap=cmap, vmin=vmin, vmax=vmax, extent=extent, alpha=alpha
+        arr_im, cmap=cmap, vmin=vmin, vmax=vmax, extent=extent, alpha=alpha, norm=norm
     )
     if title:
         ax.set(title=title)
@@ -145,6 +150,7 @@ def plot_bands(
     vmax=None,
     ax=None,
     alpha=1,
+    norm=None,
 ):
     """Plot each band in a numpy array in its own axis.
 
@@ -175,7 +181,10 @@ def plot_bands(
     alpha : float (optional)
         The alpha value for the plot. This will help adjust the transparency of the
         plot to the desired level.
-
+    norm : matplotlib Normalize object (Optional)
+        The normalized boundaries for custom values coloring. NOTE: For this arguement
+        to work, the scale arguement MUST be set to false. Otherwise, the values will
+        just be scaled from 0-255
 
     Returns
     ----------
@@ -252,6 +261,7 @@ def plot_bands(
                 title=the_title,
                 ax=ax,
                 alpha=alpha,
+                norm=norm,
             )
         # This loop clears out the plots for axes which are empty
         # A matplotlib axis grid is always uniform with x cols and x rows
@@ -285,6 +295,7 @@ def plot_bands(
             title=title,
             ax=ax,
             alpha=alpha,
+            norm=norm,
         )
         if show:
             plt.show()
