### 2019-09-12 a460c59
Made it so scale cannot be true if norm arguement is sent. Also reformatted files with black
