+++ earthpy/plot.py
@@ -82,7 +82,7 @@ def _plot_image(
     vmax=None,
     ax=None,
     alpha=1,
-    norm=None
+    norm=None,
 ):
 
     """
@@ -112,8 +112,8 @@ def _plot_image(
         The alpha value for the plot. This will help adjust the transparency of the
         plot to the desired level.
     norm : matplotlib Normalize object (Optional)
-        The normalized boundaries for custom values coloring. NOTE: For this arguement
-        to work, the scale arguement MUST be set to false. Otherwise, the values will
+        The normalized boundaries for custom values coloring. NOTE: For this argument
+        to work, the scale argument MUST be set to false. Otherwise, the values will
         just be scaled from 0-255
         
     Returns
@@ -126,7 +126,13 @@ def _plot_image(
         arr_im = es.bytescale(arr_im)
 
     im = ax.imshow(
-        arr_im, cmap=cmap, vmin=vmin, vmax=vmax, extent=extent, alpha=alpha, norm=norm
+        arr_im,
+        cmap=cmap,
+        vmin=vmin,
+        vmax=vmax,
+        extent=extent,
+        alpha=alpha,
+        norm=norm,
     )
     if title:
         ax.set(title=title)
@@ -183,8 +189,9 @@ def plot_bands(
         plot to the desired level.
     norm : matplotlib Normalize object (Optional)
         The normalized boundaries for custom values coloring. NOTE: For this arguement
-        to work, the scale arguement MUST be set to false. Otherwise, the values will
-        just be scaled from 0-255
+        to work, the scale arguement MUST be set to false. Because of this, the 
+        function will automatically set scale to false, even if the user manually
+        sets scale to true. 
 
     Returns
     ----------
@@ -211,7 +218,8 @@ def plot_bands(
         arr.ndim
     except AttributeError:
         raise AttributeError("Input arr should be a numpy array")
-
+    if norm:
+        scale = False
     if title:
         if isinstance(title, str):
             title = [title]
