+++ earthpy/plot.py
@@ -188,8 +188,8 @@ def plot_bands(
         The alpha value for the plot. This will help adjust the transparency of the
         plot to the desired level.
     norm : matplotlib Normalize object (Optional)
-        The normalized boundaries for custom values coloring. NOTE: For this arguement
-        to work, the scale arguement MUST be set to false. Because of this, the 
+        The normalized boundaries for custom values coloring. NOTE: For this argument
+        to work, the scale argument MUST be set to false. Because of this, the 
         function will automatically set scale to false, even if the user manually
         sets scale to true. 
 
