### 2019-10-01 #3 @lwasser
hey team. i thijnk we might need to regroup on this review! given there has been no activity here. i suggest we reassign everything: editor, and reviewers. @luizirber i feel like you are super busy and don't have time. i'm wondering if @marskar or someone else might want to serve as the editor and we can find two new reviewers then! i have one reviewer in mind who's using this package for lessons for the carpentries!! let me know what you guys think. a package review shouldn't sit for this long!

many thanks!
