### 2019-10-01 #3 @marskar
Hi @lwasser, I am happy to step in to serve as editor. In my opinion, *the more, the merrier* in terms of reviewers. Earth science is not my field, though I am happy to learn. It would be excellent if we could have involve the reviewer you mentioned, the one with experience using EarthPy!
