### 2019-10-01 #3 @xmnlab
I don't have any experience with EarthPy but I have some experience with GIS ... let me know if I can help in any way.
