### 2019-10-01 #3 @lwasser
thank you @xmnlab because earthpy is something i worked on, i'm trying hard to stand back from the review :) but it is a spatial library! it's designed to make exploring spatial data a bit easier. it has wrappers around rasterio and geopandas. you just did a (great) review and i know you are wrapping that up. i'll step back and will let @marskar decide on what he'd like to do as editor :) @marskar if you have any questions about the review / editor process please do say the word as well. we can also chat on thursday. Thank you both for stepping up. i think people are just very busy and i totally get that!!
