### 2019-10-04 #3 @HaoZeke
Yup, @luizirber and @lwasser, I'd be happy to review this one here and later for JOSS once it's been written up and submitted there as well.
