### 2019-10-04 #3 @luizirber
Awesome, thanks @HaoZeke!

So let's get this train rolling: @HaoZeke and @marskar have been assigned, next steps are:

- Post [this template](https://www.pyopensci.org/dev_guide/appendices/templates#review-template) in a comment in this issue
- Follow the template and https://www.pyopensci.org/dev_guide/peer_review/reviewer_guide.html for you review.
