### 2019-10-08 #3 @lwasser
hi colleagues. whoever does review this please note that the clip module is being [migrated to geopandas (it fits there better!!)](https://github.com/geopandas/geopandas/pull/1128) it's underway. but ofcourse if you have implemention comments i'll totally move that over to geopandas in that pr!
