### 2019-10-10 #3 @luizirber
@HaoZeke and @marskar can you please post your reviewer checklist to acknowledge you started?

And @marskar, do you prefer the 3 reviewers approach or moving to another review?

Thanks!
