### 2019-10-16 #3 @luizirber
Thanks for the feedback, @marskar! Let's proceed with 2 reviewers then.

@sgillies are you still available for this review?

@HaoZeke No worries! Review due date is Oct 25th, and might be postponed depending on @sgillies response. And thanks for posting the reviewer checklist =]
