### 2019-10-16 #3 @sgillies
@luizirber I am. Shall I also copy the checklist into this issue and begin? I have some evening time this week and expect no trouble with the 25 October deadline.
