### 2019-10-17 #3 @lwasser
you guys rock!! again just a note that the CLIP module is moving to GEOPANDAS:

https://github.com/geopandas/geopandas/pull/1128

but i am open to feedback on it as we move it over if you wish to provide it :)
