### 2019-10-17 f74b647
Add norm and alpha arguments (#421)
* Added alpha arguements to plot_bands and hist

* Added issue numbers to the change

* Added alpha arguement to hist and plot_bands, and a norm arguement to plot_bands, as well as tests for both.

* Made it so scale cannot be true if norm arguement is sent. Also reformatted files with black

* I still can't spell argument correctly :(

* clarify object type
