### 2019-10-19 7b762e2
import earthpy as ep, not et
To fit with the code in the following block, yes?
