### 2019-10-22 6f4d3a2
import the plot module to support accessing the plot functions!
