### 2019-10-24 earthlab/earthpy#433 @sgillies
Explain relationship that earthpy has or doesn't have to commercial software?
Earthpy explains its relationship to lower-level open source packages like rasterio and geopandas well: it builds on top of them to make some common tasks easier and gives its user more leverage. In reviewing earthpy for https://github.com/pyOpenSci/software-review/issues/3 I wonder if earthpy needs or doesn't need to explain how it relates to commercial software widely used in the higher ed research and teaching domain? This question is not loaded, I promise!

@lwasser if you don't think this needs to be added to the README, just say so, you are much more of an expert in the teaching field, and I will tick the box in the review.
