### 2019-10-24 earthlab/earthpy#434 @sgillies
Within et.data.get_data(), consider using Python's logging instead of print()
Context: https://github.com/pyOpenSci/software-review/issues/3

When get_data() is called, text is printed to stdout.

```
>>> data_path = et.data.get_data("vignette-landsat")
Downloading from https://ndownloader.figshare.com/files/15197339
Extracted output to /Users/seang/earth-analytics/data/vignette-landsat/.
```

On the one hand, some feedback to users is a good thing. On the other hand, library functions should avoid printing to stdout and should use Python's logging facility instead so that users have more control over the output.
