### 2019-10-24 earthlab/earthpy#435 @sgillies
Consider a more specific/unique directory for tutorial data
Context: https://github.com/pyOpenSci/software-review/issues/3.

Downloaded tutorial data is saved to a `earth-analytics/data` directory in the user's home directory.

```python
>>> data_path = et.data.get_data("vignette-landsat")
Downloading from https://ndownloader.figshare.com/files/15197339
Extracted output to /Users/seang/earth-analytics/data/vignette-landsat/.
>>> data_path
'/Users/seang/earth-analytics/data/vignette-landsat/.'
```
An `~/.earthpy` directory would be more like other projects like pip, which creates a `~/.pip` directory on OS X.
