### 2019-10-24 earthlab/earthpy#436 @sgillies
Change hillshade's keyword arguments to azimuth and altitude?
Context: https://github.com/pyOpenSci/software-review/issues/3.

Could the keyword arguments of https://github.com/earthlab/earthpy/blob/16b428827415453c654cc93a366f356a42c9d3c7/earthpy/spatial.py#L559 be made more consistent by changing them to azimuth and altitude and dropping the `angle_` prefix for the latter?
