### 2019-10-24 earthlab/earthpy#437 @sgillies
Reconsider exceptions raised in validating clip_shp args
Context: https://github.com/pyOpenSci/software-review/issues/3.

In clip.py:clip_shp, the two args are validated in this way:

```python
    try:
        shp.geometry
        clip_obj.geometry
    except AttributeError:
        raise AttributeError(
            "Please make sure that your input and clip GeoDataFrames have a"
            " valid geometry column"
        )
 ```

One school of thought (see Brett Slatkin’s Effective Python) is that earthpy should raise a package-specific exception here where the error arises from improper earthpy usage. Something like

```python
if not isinstance(shp, GeoDataFrame):
    raise EarthpyTypeError(“earthpy requires shp to be a GeoDataFrame”)
if not hasattr(shp, “geometry”):
    raise EarthpyValueError(“earthpy requires shp to have a geometry column”)
```

Alternatively, raise TypeError from the first test and ValueError from the second.
