### 2019-10-24 #3 @sgillies
@luizirber the earthpy package has not also been submitted to JOSS correct? I can leave those boxes unchecked?
