### 2019-10-24 #3 @HaoZeke
Yeah I believe it is not for consideration there since there's no
`paper.md` either..
