### 2019-10-24 #3 @lwasser
@sgillies @HaoZeke we'd like to submit to JOSS but i've never been through that process before and this is my first time having a package reviewed. So i was going to complete this initial review with pyopensci and then if all goes well, ask for things to be pushed along to JOSS. i need to contact arfon again about how that works. I believe that the JOSS reviewers look at the paper.md. @luizirber do you know given your experience with JOSS?? if you need the paper now i can begin working on it!! i can ask @leouieda if you guys don't know. i just need more info :)
