### 2019-10-24 earthlab/earthpy#433 @lwasser
hi @sgillies !! this is a great question and i'd love to better understand what you are thinking!

What types of commercial software are you thinking about? Like GIS type GUI based tools ??

We modeled earthpy after the `raster` and `sp` packages in `R`. These tools made is a lot easier to handle spatial manipulations in the R language that that didn't' seem to exist for python!

As far are commercial software that may be related - i am not sure what tools I would call attention to but i'm very open to suggestions if you had any in mind!! i'm not adverse to updating the readme if it would make the goal of our package more clear. @joemcglinchy @mbjoseph @jlpalomino  can you have a look at this issue and see what you think about other packages that might put earthpy into better context? The R tools are the ones that come to mind for me.
