### 2019-10-24 #3 @HaoZeke
Well I've reviewed seven papers for joss and if you set up a paper.md I'd
be happy to serve as a reviewer when you submit there as well. I'm not sure
how the transfer works though.

via Newton Mail
[https://cloudmagic.com/k/d/mailapp?ct=pa&cv=10.0.25&pv=6.0.1&source=email_footer_2]
