### 2019-10-24 earthlab/earthpy#436 @lwasser
totally @sgillies we can do that!! i'm going to allocate some time today and tomorrow to work on these issues!!
