### 2019-10-24 #3 @lwasser
ok thank you @HaoZeke i will dig into the joss part. This is the first package that will move on to JOSS. it needs to pass pyopensci review first however. I will do a bit of research on that document and will get back to you guys soon!! in the meantime i'm working on comments and issues from this review and are so appreciative of everything suggested so far!!
