### 2019-10-25 earthlab/earthpy#433 @sgillies
@lwasser I think it would be very good if the R packages this is modeled after were mentioned, and if I read the intent of the package correctly, that earthpy is intended for interactive use in a notebook. If earthpy wasn't meant to replace or integrate with commercial software, then no need to mention any.
