### 2019-10-25 earthlab/earthpy#434 @sgillies
That said, I scanned the package and this is the only use of print that I could find. If earthpy is mostly for interactive active use in a notebook and not for developing non-interactive programs, calling print from get_data is harmless. Feel free to close this one.
