### 2019-10-25 earthlab/earthpy#433 @mbjoseph
Hi @sgillies and @lwasser! Some thoughts here:

- I agree we should mention the R packages that inspired EarthPy. Both `raster` and `sp` are wrappers around lower level libraries, but provide a higher level of abstraction, which makes them a bit less flexible but perhaps easier to use for beginners. EarthPy is similar in this sense.
- We tend to use EarthPy in notebooks (in the classroom), but it will work in scripts as well. We don't (AFAIK) have any functionality that is directly tied to a notebook execution environment.
- EarthPy isn't intended to integrate with or replace any particular commercial GIS software in general, so I'd lean towards not mentioning it.

Thanks @sgillies for a thoughtful/comprehensive review! We really appreciate it.
