### 2019-10-25 2472486
Modify Hillshade argument names to add consistency (#438)
* Changed variable name in hillshade to add consistency

* Fixed tests to match new variable names

* Fixed the hillshade vignette
