### 2019-10-25 earthlab/earthpy#436 @lwasser
this has been implemented @sgillies !! closing but please reopen if our implementation is not what you intended!! #438 addresses this.
