### 2019-10-25 309d93f
update readme with preference to install via conda
