+++ README.md
@@ -17,46 +17,38 @@ EarthPy is makes it easier to plot and manipulate spatial data in Python.
 ## Why EarthPy?
 
 Python is a generic programming language designed to support many different applications. Because of this, many commonly
-performed spatial tasks for science including plotting and working with spatial data take many steps of code. EarthPy 
-takes advantage of functionality developed for raster data (rasterio) and vector data (geopandas) and simplifies the 
-code needed to :
+performed spatial tasks for science including plotting and working with spatial data take many steps of code. EarthPy
+takes advantage of functionality developed for raster data (rasterio) and vector data (geopandas) and simplifies the
+code needed to:
 
 * [Stack and crop raster bands from data such as Landsat into an easy to use numpy array](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_raster_stack_crop.html)
 * [Work with masks to set bad pixels such a those covered by clouds and cloud-shadows to NA (`mask_pixels()`)](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_stack_masks.html#sphx-glr-gallery-vignettes-plot-stack-masks-py)
 * [Plot rgb (color), color infrared and other 3 band combination images (`plot_rgb()`)](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_rgb.html)
 * [Plot bands of a raster quickly using `plot_bands()`](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_bands_functionality.html)
-* [Plot histograms for a set of raster files.](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_hist_functionality.html) 
+* [Plot histograms for a set of raster files.](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_hist_functionality.html)
 * [Create discrete (categorical) legends](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_draw_legend_docs.html)
 * [Calculate vegetation indices such as Normalized Difference Vegetation Index (`normalized_diff()`)](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_calculate_classify_ndvi.html)
 * [Create hillshade from a DEM](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_dem_hillshade.html)
 * [Clip point, line, and polygon geometries](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/plot_clip.html)
 
-EarthPy also has an io module that allows users to 
+EarthPy also has an `io` module that allows users to
 
-1. Quickly access pre-created datasubsets used in the earth-analytics courses hosted 
-on [www.earthdatascience.org](https://www.earthdatascience.org) 
+1. Quickly access pre-created data subsets used in the earth-analytics courses hosted
+on [www.earthdatascience.org](https://www.earthdatascience.org)
 2. Download other datasets that they may want to use in their workflows.
 
 ## View Example EarthPy Applications in Our Documentation Gallery  
 
-Check out our [vignette gallery](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html) for 
-applied examples of using EarthPy in common spatial workflows. 
+Check out our [vignette gallery](https://earthpy.readthedocs.io/en/latest/gallery_vignettes/index.html) for
+applied examples of using EarthPy in common spatial workflows.
 
 ## Install
 
-To install, use `pip` or `conda-forge`. We encourage you to use `conda-forge` if you are a conda users. 
+EarthPy can be installed using `pip`, but we **strongly** recommend that you install it using conda and the `conda-forge` channel.
 
[…]
