### 2019-10-25 61dbedf
add sean as a contributor
