### 2019-10-25 earthlab/earthpy#433 @lwasser
ok i have a PR open and i'll add a bit more text to the readme about the inspiration for earthpy as it aligns with the R packages!!
