### 2019-10-25 943e930
add statement about earthpy design
