### 2019-10-25 earthlab/earthpy#434 @lwasser
hey @sgillies yes we mostly intend get_data to be used in an interactive environment.  so perhaps we shouldn't modify this now but note that in the future if users want to use it more in a non interactive environment, we may want to get rid of the print statements?

the other option is to set `verbose=False` for non interactive applications? is that bad practice?
