### 2019-10-25 earthlab/earthpy#435 @lwasser
hey there @sgillies !!! all of this feedback is so helpful.
In this particular case, we have a teaching application where we are trying to teach students to manage data on their computers. so we have everyone setup a structure that is `earth-analytics/data`. we also have a lesson portal (earthdatascience.org) where lots of people go and work through our lessons... having  `get_data` create the intended structure for students avoids many challenges that they face of creating paths!

we like to teach about expressive names and keeping data together in a data dir and such. given that explanation, does that structure make a bit more sense? i'm a little concerned about changing this particular item given our teaching applications. Please let me know what you think!!

again I am so appreciative of your input and review!!
