### 2019-10-25 earthlab/earthpy#437 @lwasser
oh interesting!! We are submitting this clip module now to geopandas: https://github.com/geopandas/geopandas/pull/1128 so i think we should definitely consider this in that PR!! we will remove it from here once they merge but i and @nkorinek are still working on the pr.

I will make sure this is addressed in that PR however so will leave this open for the time being!
