### 2019-10-25 #3 @lwasser
ok i am working my way through @sgillies comments and issues which are all great -- THANK YOU!! @HaoZeke do you have a review for us coming? i didn't see your review template as filled out.

in the meantime we will have a paper.md file pushed in the next few days for JOSS!!
