### 2019-10-25 5bbea11
add initial draft of paper.md to github
This is the paper required for submission to JOSS.
