+++ paper.md
@@ -1,0 +1,110 @@
+
+---
+title: EarthPy: A Python package that makes it easier to explore and plot raster and vector data using open source Python tools.'
+tags:
+  - Python
+  - gis
+  - raster data
+  - vector data
+  - remote sensing
+authors:
+  - name: Leah Wasser 
+    orcid: 0000-0002-8177-6550
+    affiliation: "Earth Lab, University of Colorado - Boulder" # (Multiple affiliations must be quoted)
+  - name: Maxwell B. Joseph
+    orcid: 0000-0002-7745-9990
+    affiliation: "Earth Lab, University of Colorado - Boulder"
+  - name: Korinek, Nathan
+    orcid: 0000-0003-0859-7246
+    affiliation: "Earth Lab, University of Colorado - Boulder"
+  - name: Jenny Palomino 
+    orcid: 0000-0003-4879-9299
+    affiliation: "Earth Lab, University of Colorado - Boulder"
+  - name: Chris Holdgraf 
+    orcid: 0000-0002-2391-0678
+    affiliation: "Berkely, Project Jupyter"
+date: 23 October 2019
+bibliography: paper.bib??
+# Optional fields if submitting to a AAS journal too, see this blog post:
+# https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing
+# We have a zenodo doi -- how does JOSS handle this?
+---
+
+# Summary & Purpose
+
+`EarthPy` makes commonly performed spatial data exploration tasks easier by building upon commonly used functionality available in the widely used packages: `Rasterio` and `GeoPandas`. `EarthPy` is designed for users who are new to `Python` and spatial data with a focus on scientific data. 
+
+When a user is working with spatial data for research, there are a suite of data exploration activities that are often performed including: 
+
+1. Viewing histograms and plots of single bands within a remote sensing image to explore potential calibration and other data quality issues.
+2. Creating basemaps that have legends with unique symbology.
+3. Creating plots of images with colorbars.
+4. Rendering RGB (and other composite) images of multi band spectral remote sensing images. 
+5. Masking clouds from a remote sensing image.
+
+The above operations are crucial to understanding a dataset and identifying issues that may need to be addressed with further data processing when beginning an analysis. In the `R` world, these tasks are quickly performed using the `raster` and `sp` packages. However, there isn't a tool that makes these tasks easy for users in the `Python` open source package landscape. 
+
+
+## EarthPy Audience 
+
+`EarthPy` was originally designed to support the Earth Analytics Education program at Earth Lab - University of Colorado, Boulder. Our program teaches students how to work with a suite of earth and environmental data using open source `Python`. All lessons are published as free open education resources on our online learning portal (https://www.earthdatascience.org). Through this publication process, we identified that suites of spatial data exploration and cleanup tasks that were performed regularly required many steps that could be easily wrapped into helper functions. We modeled these functions after those available in the `R` ecosystem given the experience of many years of teaching with `R`. 
+
+# EarthPy Functionality
+
+`EarthPy` is organized into several modules: 
+
+* [io: Input/output for data](https://earthpy.readthedocs.io/en/latest/api/earthpy.io.html): utility functions to download existing teaching data subsets or other data into a user's working directory (by default, this directory is: `~/earth-analytics/data`). The IO module supports downloading data for the Earth Lab Earth Analytics courses as well as any user with a URL to a compressed file.
+* [mask: Mask out cloud and shadow covered pixels from raster data](https://earthpy.readthedocs.io/en/latest/api/earthpy.mask.html): helper functions to mask remote sensing images using a cloud mask or QA layer. 
[…]
