### 2019-10-26 earthlab/earthpy#443 @HaoZeke
[pyOpenSci] Citation details
# Citation
This is part of https://github.com/pyOpenSci/software-review/issues/3.
I note that the code currently has a zenodo doi. So it would be pertinent to add a bibtex formatted citation in the readme.
