### 2019-10-26 earthlab/earthpy#444 @HaoZeke
[pyOpenSci] Contributing Guideline Formatting
# Harmonize Format
This issue is part of the ongoing review process https://github.com/pyOpenSci/software-review/issues/3.
Parts of the `contributing.rst` file appear to be in markdown syntax. Please fix this.

![image](https://user-images.githubusercontent.com/4336207/67623474-34922100-f843-11e9-9275-b74d31b1bfdd.png)
