### 2019-10-27 earthlab/earthpy#434 @sgillies
The option is good, but logging (where the default stream is stderr) would be even better. The situation that we want to prevent is where someone writes a script with earthpy that prints data to stdout (used like `python myscript.py > data.geojson`) and it has data download messages interleaved with the JSON text output.
