### 2019-10-27 earthlab/earthpy#435 @sgillies
@lwasser showing students how to organize their data is a great goal. I only wanted to point out that ~/.earthpy is a bit more safe from being clobbered by other software, security by obscurity. It could be too obscure, though. I don't think this issue is a blocker at all, so feel free to close it up if you like.
