### 2019-10-28 earthlab/earthpy#435 @lwasser
got ya @sgillies i appreciate this. i think for this particular application in our classes it might be best to stick with that path. but i hear you on the safety issue and will definitely consider it!! if you are ok with my closing this particular issue, i will. again i'm so thankful for all of this input. i'm learning a lot!!
