### 2019-10-28 8465db6
cleanup of markdown --> rst
