+++ CONTRIBUTING.rst
@@ -277,9 +277,10 @@ Once that is done, create a release on GitHub for the new version.
 Data Generated for Testing
 ==========================
 If a test requires a data object such as a GeoDataFrame or numpy array, and
-copies of that data object are required by multiple tests, we can use [pytest
-fixtures](https://docs.pytest.org/en/latest/fixture.html) to cleanly create
-and tear down those objects independently for each test.
-See [`earthpy/tests/conftest.py`](earthpy/tests/conftest.py) for fixture
-definitions, and [`earthpy/tests/test_clip.py`](earthpy/tests/test_clip.py)
+copies of that data object are required by multiple tests, we can use
+`pytest fixtures <https://docs.pytest.org/en/latest/fixture.html>`_ to cleanly 
+create and tear down those objects independently for each test.
+See `earthpy/tests/conftest.py <https://www.github.com/earthpy/tests/conftest.py>`_
+for fixture definitions, and 
+`earthpy/tests/test_clip.py <https://www.github.com/earthpy/earthpy/tests/test_clip.py>`_
 for example usage of fixtures in tests.
