### 2019-10-28 09eda25
Update per black (#425)
