### 2019-10-28 d36f057
add logging if verbos is set to false
