### 2019-10-28 earthlab/earthpy#434 @lwasser
thank you again, @sgillies ok i did some reading on this and started a PR

https://github.com/earthlab/earthpy/pull/448

does this approach seem reasonable? OR what would you implement here? in this case i printed if `verbose=True` (which is a default) but used `logging.info` if it's false. Please know this particular concept is new to me SO if i am implementing it incorrectly please do say the word or suggest another approach!!!
