+++ CONTRIBUTING.rst
@@ -278,9 +278,9 @@ Data Generated for Testing
 ==========================
 If a test requires a data object such as a GeoDataFrame or numpy array, and
 copies of that data object are required by multiple tests, we can use
-`pytest fixtures <https://docs.pytest.org/en/latest/fixture.html>`_ to cleanly 
+`pytest fixtures <https://docs.pytest.org/en/latest/fixture.html>`_ to cleanly
 create and tear down those objects independently for each test.
-See `earthpy/tests/conftest.py <https://www.github.com/earthpy/tests/conftest.py>`_
-for fixture definitions, and 
-`earthpy/tests/test_clip.py <https://www.github.com/earthpy/earthpy/tests/test_clip.py>`_
+See `earthpy/tests/conftest.py <https://github.com/earthlab/earthpy/blob/master/earthpy/tests/conftest.py>`_
+for fixture definitions, and
+`earthpy/tests/test_clip.py <https://github.com/earthlab/earthpy/blob/master/earthpy/tests/test_crop_image.py>`_
 for example usage of fixtures in tests.
