### 2019-10-28 554b22f
add logger object
