### 2019-10-28 58cb7a2
cleanup pr with readme and contributing edits
