### 2019-10-28 b6b42fa
small fix in changelog
