+++ CHANGELOG.md
@@ -7,12 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.7.4]
+## [0.7.5]
 * Changed variable name for `angle_altitude` to `altitude` in `hillshade()` (@nkorinek, #436)
 * Added alpha arguments to `hist()` and `plot_bands()` (@nkorinek, #409, #410)
 * Add vignette to download data from URLs and EarthPy using `data.get_data()` function (@jlpalomino, #396)
 
-## [0.7.3]
+## [0.7.4]
 * Added alpha arguments to `hist()` and `plot_bands()` (@nkorinek, #409, #410)
 
 ## [0.7.3]
