### 2019-10-28 cfe5a17
update contributing with zenodo info
