### 2019-10-28 e90e010
provide string for logger instantiation
