### 2019-10-28 50a746b
logger named after module name
