### 2019-10-28 a94d988
update affiliation re: joss standards
