### 2019-10-28 bf2b23f
more syntax fixes
