### 2019-10-28 3d2dda1
cleanup pr with readme and contributing edits following pyopensci review (#449)
* cleanup pr with readme and contributing edits

* update contributing with zenodo info

* sure heading consistently
