### 2019-10-28 earthlab/earthpy#444 @lwasser
@HaoZeke this was implemented here: https://github.com/earthlab/earthpy/pull/449 !! closing but feel free to reopen if you are not satisfied with the changes!!
