### 2019-10-28 earthlab/earthpy#443 @lwasser
i've added a link to the bibtext format in the readme!! i didn't want to copy the citation verbatim in case someone wanted to cite a particular version. please let me know what you think. closing but please reopen if you are not satisfied @HaoZeke !!!
