### 2019-10-28 earthlab/earthpy#433 @lwasser
this has been merged!! if you are not satisfied with the changes @sgillies please reopen this issue or i'm happy to do so and work on it a bit more!!
