### 2019-10-29 #3 @HaoZeke
Hi @luizirber. I have completed my review, and I am satisfied with the code both in terms of functionality as well as documentation and reproducibility. I am not much of a geologist, but I am certain such a well crafted tool will be of much use to the community. I recommend this for endorsement. Also @lwasser, thank you for your patience. I would be willing to review this submission when you submit it to JOSS as well.
