### 2019-10-29 0d0ba72
Update earthpy/io.py
Co-Authored-By: Sean Gillies <sean.gillies@gmail.com>
