### 2019-10-29 earthlab/earthpy#434 @sgillies
@nkorinek yes, changing the log level within the function is a good approach. Instead of configuring the root logger, I think the project should set the level of the module's logger `logging.getLogger(__name__).setLevel(logging.INFO)`. This will increase earthpy's verbosity only and not add noise from requests or other modules.
