### 2019-10-29 earthlab/earthpy#434 @nkorinek
@sgillies that looks much cleaner, cool!
