### 2019-10-29 f920d5a
Update earthpy/io.py
Co-Authored-By: Sean Gillies <sean.gillies@gmail.com>
