### 2019-10-29 330f39c
Update earthpy/io.py
Co-Authored-By: Sean Gillies <sean.gillies@gmail.com>
