### 2019-10-29 fd3bec0
Update earthpy/io.py
Co-Authored-By: Sean Gillies <sean.gillies@gmail.com>
