### 2019-10-31 1573f9e
remove if else and use logging at info level to print
