### 2019-10-31 d881a9f
Set default logger level to warning
