### 2019-10-31 9ee383a
fix references and minor edits
