+++ paper.md
@@ -32,12 +32,12 @@ affiliations:
   - name: "University of California - Berkeley, Project Jupyter"
     index: 2
 date: 28 October 2019
-bibliography:
+bibliography: paper.bib
 ---
 
 # Summary & Purpose
 
-`EarthPy` makes commonly performed spatial data exploration tasks easier by building upon commonly used functionality available in the widely used packages: `Rasterio` and `GeoPandas`. `EarthPy` is designed for users who are new to `Python` and spatial data with a focus on scientific data. 
+`EarthPy` makes commonly performed spatial data exploration tasks easier for scientist by building upon functions in the widely used packages: `Rasterio` and `GeoPandas`. `EarthPy` is designed for users who are new to `Python` and spatial data with a focus on scientific data. 
 
 When a user is working with spatial data for research, there are a suite of data exploration activities that are often performed including: 
 
@@ -67,7 +67,6 @@ The above operations are crucial to understanding a dataset and identifying issu
 * [spatial: Raster processing and analysis](https://earthpy.readthedocs.io/en/latest/api/earthpy.spatial.html): utilities to crop a set of bands to a defined spatial extent, create a hillshade, stack bands, and calculate normalized difference rasters.
 * [clip: Vector data subsetting](https://earthpy.readthedocs.io/en/latest/api/earthpy.clip.html): A module to clip vector data using GeoPandas. Allows for clipping of points, lines, and polygon data within a specified polygon.
 
-
 ## EarthPy Vignettes 
 
 In addition to detailed API documentation and example code executed by doctest, 
@@ -84,7 +83,7 @@ convey key concepts for spatial data processing.
 
 `EarthPy` is an open source `Python` package that makes it easier to plot and work with both spatial raster and vector data using open source tools. `EarthPy`'s goal is to make working with spatial data easier for scientists who want to use open source `Python` tools for analysis and visualization. 
 
-`Earthpy` depends upon `GeoPandas`, which has a focus on vector data handling and analysis, and `Rasterio`, which facilitates input and output of raster data files as numpy arrays. It also requires `Matplotlib` for plotting operations. 
+`Earthpy` depends upon `GeoPandas` [@kelsey_jordahl_2019_3483425:2019], which has a focus on vector data handling and analysis, and `Rasterio` [@gillies:2013], which facilitates input and output of raster data files as numpy arrays. It also requires `Matplotlib` for plotting operations. 
 
 To simplify dependency management and installation for non-experts, we maintain a version of EarthPy on the `conda-forge` channel, which installs the system 
 libraries upon which `EarthPy` depends. This combined with high-level wrapper 
@@ -103,15 +102,8 @@ The `plot` module facilitates quick and early data exploration by introductory-l
 
 The vignettes developed with `EarthPy` also provide easily adaptable starting points for in-class exercises that help students learn key spatial concepts using scientific data. 
 
-
-## References
-
-* [Rasterio](https://rasterio.readthedocs.io/en/stable/intro.html)
-* [GeoPandas](http://geopandas.org/)
-* [PySAL](http://pysal.org/pysal/)
-* [GeoRasters](https://github.com/ozak/georasters)
-
-
-## Acknowledgements
+# Acknowledgements
 
 There have been many [contributors to earthpy](https://github.com/earthlab/earthpy/graphs/contributors) that we are thankful for. We are also thankful for the feedback that we recieved through the software review implemented by pyOpenSci. Specifically we thank Luiz Irber who has served as an editor for this review and the two reviewers: Sean Gillies and Rohit Goswami. 
+
+# References
