### 2019-10-31 a03e40d
add bib files
