+++ paper.bib
@@ -1,0 +1,48 @@
+@software{kelsey_jordahl_2019_3483425,
+  author       = {Kelsey Jordahl and
+                  Joris Van den Bossche and
+                  Jacob Wasserman and
+                  James McBride and
+                  Jeffrey Gerard and
+                  Martin Fleischmann and
+                  Jeff Tratner and
+                  Matthew Perry and
+                  Carson Farmer and
+                  Geir Arne Hjelle and
+                  Sean Gillies and
+                  Micah Cochran and
+                  Matt Bartos and
+                  Lucas Culbertson and
+                  Nick Eubank and
+                  maxalbert and
+                  Sergio Rey and
+                  Aleksey Bilogur and
+                  Dani Arribas-Bel and
+                  Christopher Ren and
+                  Joshua Wilson and
+                  Martin Journois and
+                  Levi John Wolf and
+                  Leah Wasser and
+                  Ömer Özak and
+                  YuichiNotoya and
+                  François Leblanc and
+                  Filipe and
+                  Chris Holdgraf and
+                  Adam Greenhall},
+  title        = {geopandas/geopandas: v0.6.1},
+  month        = oct,
+  year         = 2019,
+  publisher    = {Zenodo},
+  version      = {v0.6.1},
+  doi          = {10.5281/zenodo.3483425},
+  url          = {https://doi.org/10.5281/zenodo.3483425}
+}
+
+
+@Misc{gillies,
+  author =    {Sean Gillies and others},
+  organization = {Mapbox},
+  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
+  year =      {2013--},
+  url = "https://github.com/mapbox/rasterio"
+}
