### 2019-10-31 a266ef4
Add paper.md to support JOSS review (#441)
* add initial draft of paper.md to github

This is the paper required for submission to JOSS.

* more updates from joe!!

* update affiliation re: joss standards

* update affiliation re: joss standards

* more syntax fixes

* more syntax fixes

* whedon be happy

* fix references and minor edits

* add bib files
