### 2019-10-31 65e3c79
only return paths if verbose is true
