### 2019-10-31 1f4b0de
Test that the logs are present and at the correct level (#451)
