+++ paper.bib
@@ -39,7 +39,7 @@
 }
 
 
-@Misc{gillies,
+@software{gillies_2019,
   author =    {Sean Gillies and others},
   organization = {Mapbox},
   title =     {Rasterio: geospatial raster I/O for {Python} programmers},
