### 2019-11-01 fb069ff
updating paper with new citation for rasterio
