### 2019-11-02 #3 @luizirber
Hello all,

thanks for the reviews @sgillies and @HaoZeke! I see that the only item missing in both your checklist is the final approval (and the JOSS parts), so I think this is ready to be approved (if you don't have any other concerns).
