### 2019-11-04 #3 @lwasser
awesome. just a note that
1. we will not address the issue associated with the clip modules as that is moving to geopandas and we are working on the PR there. we will however integrate sean's suggestions in that PR!
2. we were going to implement logging but after some discussion with @sgillies we decided that this might not be necessary for earthpy. so if everyone is on board, i think we are in good shape. I do have the joss paper in the repo now AND it's passing all of the JOSS whedon tests. so when i get the final approval here, i'll submit it to JOSS. i am more than happy however to address any other items if need be !!!
