### 2019-11-04 #3 @sgillies
@lwasser went above and beyond in responding to my review comments. I've checked off the last box. Thanks for letting me get involved, I found the process and documentation quite interesting and well done.
