### 2019-11-04 earthlab/earthpy#434 @lwasser
ok thank you again for this suggestion @sgillies  for the time being i'm going to close this and the associated PR. but perhaps we will revisit it in the future if it becomes a real problem for people using the package!! your input and review has been greatly appreciated!! THANK YOU!
