### 2019-11-06 #3 @luizirber
Thanks @sgillies!

@HaoZeke, any final comments before we move to `approved`?
