### 2019-11-06 #3 @HaoZeke
Thank you for asking. I think the documentation is great and it's ready to
go further without any issues. Great work.
