### 2019-11-06 #3 @luizirber
Thanks @HaoZeke!

Label updated, now EarthPy is [`6/approved`](https://github.com/pyOpenSci/software-review/labels/6%2Fapproved)!
