### 2019-11-06 #3 @luizirber
and this approval template might need some tweaks, since we are not asking repos to be moved to pyopensci... But the JOSS items are relevant!
