### 2019-11-06 #3 @lwasser
oh yay !! ok @luizirber that's a great idea to update the template. whatever suggestions you have would be awesome!!

I will create a release and submit to JOSS next. @arfon told me to just link to this issue as documentation that we are approved by pyopensci. You may know more than me on this but i'll submit it first and will report back here in this issue. Please let me know if you have any suggestions!!

@HaoZeke  and @sgillies thank you AGAIN for your reviews. i think we made some great improvements to earthpy through this process.
