### 2019-11-07 38ffa30
FIX TWO CITATIONS that were causing JOSS build errors
