### 2019-11-07 #3 @lwasser
JOSS review is happening here: https://github.com/openjournals/joss-reviews/issues/1869
