### 2019-11-07 cd23b6f
fix black and also badge order on readme
