### 2019-11-07 37867b0
don't run docs on 3.7
