### 2019-11-07 b3dabda
print log
