+++ .travis.yml
@@ -26,7 +26,7 @@ script:
   - ./.travis/test.sh
 
 after_failure:
-  - cat MSBuild_*.failure.txt
+  - cat /home/travis/build/earthlab/earthpy/.tox/black/log/black-2.log
 
 deploy:
   provider: pypi
