+++ tox.ini
@@ -19,6 +19,8 @@ commands =
 [testenv:black]
 deps = black
 basepython = python3
+deps =
+    -r{toxinidir}/dev-requirements.txt
 commands = black --check --verbose earthpy
 
 [testenv:docs]
