+++ tox.ini
@@ -19,8 +19,10 @@ commands =
 [testenv:black]
 deps = black
 basepython = python3
+#deps =
+#    -r{toxinidir}/dev-requirements.txt
 deps =
-    -r{toxinidir}/dev-requirements.txt
+    pip>=19.0
 commands = black --check --verbose earthpy
 
 [testenv:docs]
