+++ tox.ini
@@ -17,12 +17,9 @@ commands =
     codecov -e TOXENV
 
 [testenv:black]
-deps = black
-basepython = python3
-#deps =
-#    -r{toxinidir}/dev-requirements.txt
 deps =
-    pip>=19.0
+  pip>=19.0, black
+basepython = python3
 commands = black --check --verbose earthpy
 
 [testenv:docs]
