+++ tox.ini
@@ -18,7 +18,8 @@ commands =
 
 [testenv:black]
 deps =
-  pip>=19.0, black
+  pip>=19.0
+  black
 basepython = python3
 commands = black --check --verbose earthpy
 
