### 2019-11-07 69912bf
add docs back to build
