+++ .travis.yml
@@ -13,7 +13,7 @@ matrix:
         - os: linux
           dist: xenial
           python: 3.7
-          env: TOXENV=py37,black #,docs
+          env: TOXENV=py37, black ,docs
         - os: osx
           language: generic
 
