### 2019-11-07 1fd95f4
fix black and also badge order on readme (#452)
* fix black and also badge order on readme

* potential pip fix for pyproj

* don't run docs on 3.7

* print log

* add black back

* specify envt for black

* force pip verion in black

* specify pip in deps

* another try at deps install

* add docs back to build
