### 2019-11-11 ed82c38
add citation for matplotlib
