### 2019-11-11 8c73bdc
adding pysal citation
still can't find one for georasters
