+++ paper.bib
@@ -1,3 +1,26 @@
+@software{gillies_2019,
+  author =    {Sean Gillies and others},
+  organization = {Mapbox},
+  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
+  year =      {2013--},
+  url = "https://github.com/mapbox/rasterio"
+}
+
+@Article{Hunter:2007,
+  Author    = {Hunter, J. D.},
+  Title     = {Matplotlib: A 2D graphics environment},
+  Journal   = {Computing in Science \& Engineering},
+  Volume    = {9},
+  Number    = {3},
+  Pages     = {90--95},
+  abstract  = {Matplotlib is a 2D graphics package used for Python for
+  application development, interactive scripting, and publication-quality
+  image generation across user interfaces and operating systems.},
+  publisher = {IEEE COMPUTER SOC},
+  doi       = {10.1109/MCSE.2007.55},
+  year      = 2007
+}
+
 @software{kelsey_jordahl_2019_3483425,
   author       = {Kelsey Jordahl and
                   Joris Van den Bossche and
@@ -38,26 +61,13 @@
   url          = {https://doi.org/10.5281/zenodo.3483425}
 }
 
-
-@software{gillies_2019,
-  author =    {Sean Gillies and others},
-  organization = {Mapbox},
-  title =     {Rasterio: geospatial raster I/O for {Python} programmers},
-  year =      {2013--},
-  url = "https://github.com/mapbox/rasterio"
-}
-
-@Article{Hunter:2007,
-  Author    = {Hunter, J. D.},
-  Title     = {Matplotlib: A 2D graphics environment},
-  Journal   = {Computing in Science \& Engineering},
-  Volume    = {9},
-  Number    = {3},
-  Pages     = {90--95},
-  abstract  = {Matplotlib is a 2D graphics package used for Python for
-  application development, interactive scripting, and publication-quality
-  image generation across user interfaces and operating systems.},
-  publisher = {IEEE COMPUTER SOC},
-  doi       = {10.1109/MCSE.2007.55},
-  year      = 2007
+@Article{pysal2007,
+  author={Rey, Sergio J. and Anselin, Luc},
+  title={{PySAL: A Python Library of Spatial Analytical Methods}},
+  journal={The Review of Regional Studies},
+  year=2007,
+  volume={37},
+  number={1},
+  pages={5-27},
+  keywords={Open Source; Software; Spatial}
 }
