### 2019-11-13 7c8c0a7
Finalize edits for JOSS review (#455)
* adding tim as an author to align with zenodo and a few other minor syntax edits.
