+++ paper.md
@@ -8,7 +8,7 @@ tags:
   - vector data
   - remote sensing
 authors:
-  - name: Leah Wasser 
+  - name: Leah Wasser
     orcid: 0000-0002-8177-6550
     affiliation: 1
   - name: Maxwell B. Joseph
@@ -17,63 +17,68 @@ authors:
   - name: Joe McGlinchy
     orcid: 0000-0003-2135-0168
     affiliation: 1
-  - name: Jenny Palomino 
+  - name: Jenny Palomino
     orcid: 0000-0003-4879-9299
     affiliation: 1
   - name: Korinek, Nathan
     orcid: 0000-0003-0859-7246
     affiliation: 1
-  - name: Chris Holdgraf 
+  - name: Chris Holdgraf
     orcid: 0000-0002-2391-0678
     affiliation: 2
+  - name: Tim Head
+    orcid: 0000-0003-0931-3698
+    affiliation: 3
 affiliations:
   - name: "Earth Lab, University of Colorado - Boulder"
     index: 1
   - name: "University of California - Berkeley, Project Jupyter"
     index: 2
+  - name: "Wild Tree Tech"
+    index: 3
 date: 28 October 2019
 bibliography: paper.bib
 ---
 
 # Summary & Purpose
 
-`EarthPy` makes commonly performed spatial data exploration tasks easier for scientist by building upon functions in the widely used packages: `Rasterio` and `GeoPandas`. `EarthPy` is designed for users who are new to `Python` and spatial data with a focus on scientific data. 
+`EarthPy` makes commonly performed spatial data exploration tasks easier for scientist by building upon functions in the widely used packages: `Rasterio` and `GeoPandas`. `EarthPy` is designed for users who are new to `Python` and spatial data with a focus on scientific data.
 
-When a user is working with spatial data for research, there are a suite of data exploration activities that are often performed including: 
+When a user is working with spatial data for research, there are a suite of data exploration activities that are often performed including:
 
 1. Viewing histograms and plots of single bands within a remote sensing image to explore potential calibration and other data quality issues.
 2. Creating basemaps that have legends with unique symbology.
 3. Creating plots of images with colorbars.
-4. Rendering RGB (and other composite) images of multi band spectral remote sensing images. 
+4. Rendering RGB (and other composite) images of multi band spectral remote sensing images.
 5. Masking clouds from a remote sensing image.
 6. Limiting geographic extent of spatial data
 
-The above operations are crucial to understanding a dataset and identifying issues that may need to be addressed with further data processing when beginning an analysis. In the `R` world, these tasks are quickly performed using the `raster` and `sp` packages. However, there isn't a tool that makes these tasks easy for users in the `Python` open source package landscape. 
+The above operations are crucial to understanding a dataset and identifying issues that may need to be addressed with further data processing when beginning an analysis. In the `R` world, these tasks are quickly performed using the `raster` and `sp` packages. However, there isn't a tool that makes these tasks easy for users in the `Python` open source package landscape.
 
 
-## EarthPy Audience 
+## EarthPy Audience
 
-`EarthPy` was originally designed to support the Earth Analytics Education program at Earth Lab - University of Colorado, Boulder. Our program teaches students how to work with a suite of earth and environmental data using open source `Python`. All lessons are published as free open education resources on our online learning portal (https://www.earthdatascience.org). Through this publication process, we identified that many spatial data exploration and cleanup tasks which were performed regularly required many steps that could be easily wrapped into helper functions. We modeled these functions after those available in the `R` ecosystem, given the experience of the developers' many years of working and teaching with `R`. 
[…]
