+++ docs/api/earthpy.mask.rst
@@ -1,4 +1,4 @@
 .. automodule:: earthpy.mask
-    :members:
-    :undoc-members:
-    :show-inheritance:
+   :members:
+   :undoc-members:
+   :show-inheritance:
