### 2019-11-13 67ab319
Bump version: 0.7.6 → 0.8.0
