+++ setup.py
@@ -22,7 +22,7 @@ if __name__ == "__main__":
         description=DESCRIPTION,
         long_description=LONG_DESCRIPTION,
         long_description_content_type="text/markdown",
-        version="0.7.6",
+        version="0.8.0",
         packages=["earthpy"],
         install_requires=[
             "geopandas",
