### 2019-11-13 cbf12d3
add JOSS badge to earthpy!!
yay!
