### 2019-11-13 1805290
Suggested updates to README
