### 2025-02-15 #3 @banesullivan
I have just applied the [archive label](https://github.com/pyOpenSci/software-submission/issues?q=label%3Aarchived) to this review per @lwasser's request in https://github.com/pyOpenSci/pyopensci.github.io/issues/560#issuecomment-2654342122 per the guidelines in the [review guide](https://github.com/pyOpenSci/pyopensci.github.io/issues/560#issuecomment-2654342122)
